Refactor key selector dialog into package with mixin modules
- Extract KeySelectorDialog into key_selector/ package with tab and option mixins
- Split analog, macro, superkey, profile, gamepad axis, and options into separate modules
- Add compat shim module for import path migration
- Extract gamepad output helpers into reusable functions with Protocol typing
- Add PositionCaptureMixin for coordinate capture
