+++ keymasq/gui/widgets/key_selector/__init__.py
@@ -1,0 +1,4 @@
+from keymasq.gui.widgets.key_selector.dialog import KeySelectorDialog
+from keymasq.gui.widgets.key_selector.superkey_action_dialog import SuperkeyActionDialog
+
+__all__ = ["KeySelectorDialog", "SuperkeyActionDialog"]
