+++ keymasq/gui/widgets/key_selector_dialog.py
@@ -1,8 +1,5 @@
-import logging
-from collections.abc import Callable
-from typing import cast
+from __future__ import annotations
 
-import evdev
 import gi
 
 gi.require_version("Gtk", "4.0")
@@ -17,4176 +14,87 @@ from gi.repository import (  # pyright: ignore[reportAttributeAccessIssue]
 )
 
 from keymasq import __version__
-from keymasq.common.devices import resolve_evdev_code
-from keymasq.common.gamepad_axes import (
-    GAMEPAD_AXIS_RANGES,
-    gamepad_axis_percent_from_value,
-    gamepad_axis_range,
-    gamepad_axis_value_from_percent,
-    normalize_gamepad_axis_target,
-)
-from keymasq.common.models import (
-    DEFAULT_REPEAT_CATEGORIES,
-    MAX_MACRO_RECORDING_SLOTS,
-    MIN_RAPIDFIRE_HOLD_MS,
-    MIN_RAPIDFIRE_WAIT_MS,
-    REPEAT_CATEGORY_GAMEPAD,
-    REPEAT_CATEGORY_KEYBOARD,
-    REPEAT_CATEGORY_MACRO,
-    REPEAT_CATEGORY_MOUSE,
-    REPEAT_CATEGORY_SPECIAL,
-    ActionType,
-    AnalogControlConfig,
-    MappingAction,
-    ProfileDeactivationPolicy,
-    SuperkeyAction,
-    SuperkeyConfig,
-    action_type_supports_rapidfire,
-)
 from keymasq.common.slurp import get_slurp_capture
 from keymasq.gui.session_client import session_request_async
 from keymasq.gui.session_reload import notify_session_reload_async
-from keymasq.gui.widgets.action_labels import describe_mapping_action_verbose
-from keymasq.gui.widgets.compositor_actions import (
-    build_compositor_action_pages,
-    compositor_action_tab_name,
+from keymasq.gui.widgets.gamepad_output_choices import virtual_gamepad_count
+from keymasq.gui.widgets.input_picker_shared import GAMEPAD_BUTTONS
+from keymasq.gui.widgets.key_selector import KeySelectorDialog, SuperkeyActionDialog
+from keymasq.gui.widgets.key_selector.gamepad_axis import (
+    GamepadAxisControlsMixin as _GamepadAxisControlsMixin,
 )
-from keymasq.gui.widgets.docs_links import actions_docs_url
-from keymasq.gui.widgets.fuzzy_search import (
-    fuzzy_query_matches,
-    install_listbox_fuzzy_filter,
-    macro_search_text,
+from keymasq.gui.widgets.key_selector.tabs import (
+    _create_actions_docs_button,
+    _ensure_compact_tabs_css,
 )
-from keymasq.gui.widgets.gamepad_output_choices import (
-    gamepad_output_choice_matches,
-    gamepad_output_choices_for,
-    gamepad_output_unavailable_message,
-    virtual_gamepad_count,
-)
-from keymasq.gui.widgets.input_picker_shared import (
-    GAMEPAD_BUTTONS,
-)
-from keymasq.gui.widgets.input_picker_shared import (
-    build_gamepad_tab as build_shared_gamepad_tab,
-)
-from keymasq.gui.widgets.input_picker_shared import (
-    build_keyboard_tab as build_shared_keyboard_tab,
-)
-from keymasq.gui.widgets.input_picker_shared import (
-    build_media_tab as build_shared_media_tab,
-)
-from keymasq.gui.widgets.input_picker_shared import (
-    build_mouse_tab as build_shared_mouse_tab,
-)
-from keymasq.gui.widgets.input_picker_shared import (
-    build_navigation_tab as build_shared_navigation_tab,
-)
-from keymasq.session.analog_controls import (
-    AnalogControlManager,
-    AnalogControlPreset,
-    analog_control_presets,
+from keymasq.gui.widgets.key_selector.targets import (
+    _GAMEPAD_AXIS_CUSTOM_SLOT,
+    _PROFILE_LIFETIME_PRESETS_ENABLE,
+    _PROFILE_LIFETIME_PRESETS_TOGGLE,
+    ACTION_DOC_LINKS,
+    DEFAULT_RAPIDFIRE_TOOLTIP,
+    EVDEV_TO_GAMEPAD,
+    EVDEV_TO_KEY,
+    F_EXTRA,
+    KEY_TO_EVDEV,
+    KEY_WIDTHS,
+    KEYBOARD_LAYOUT,
+    MEDIA_KEY_GROUPS,
+    MEDIA_KEY_TARGETS,
+    MPRIS_MEDIA_GROUPS,
+    REPEAT_CATEGORY_OPTIONS,
+    REPEAT_RAPIDFIRE_TOOLTIP,
+    SYSTEM_KEY_GROUPS,
+    SYSTEM_KEY_TARGETS,
+    _actions_docs_url,
+    _keyboard_target_allows_rapidfire,
+    _keyboard_target_allows_tap,
+    _resolve_gamepad_axis_target,
+    _resolve_gamepad_button_target,
 )
 from keymasq.session.compositor import detect_compositor_sync
 from keymasq.session.hardware import HardwareManager
-from keymasq.session.profiles import ProfileManager
-from keymasq.session.superkeys import SuperkeyManager
 
-log = logging.getLogger("keymasq.gui.widgets.key_selector_dialog")
-
-_PROFILE_LIFETIME_PRESETS_ENABLE: tuple[tuple[str, str], ...] = (
-    ("until_changed", "Persistent"),
-    ("while_trigger_active", "While trigger is held"),
-    ("after_one_action", "One-shot"),
-    ("custom", "Custom"),
-)
-_PROFILE_LIFETIME_PRESETS_TOGGLE: tuple[tuple[str, str], ...] = (
-    ("until_changed", "Persistent"),
-    ("after_one_action", "One-shot"),
-    ("custom", "Custom"),
-)
-
-
-def _resolve_gamepad_button_target(raw: str) -> str | None:
-    """Resolve free-form button input to a canonical evdev button name.
-
-    Accepts an evdev button name (e.g. ``btn_c``, ``btn_trigger_happy1``) or a
-    numeric key code (decimal or ``0x``-prefixed). Returns a lowercase ``btn_*``
-    name, or ``None`` when the input is not a recognized button code.
-    """
-    text = (raw or "").strip().lower()
-    if not text:
-        return None
-
-    if text.startswith("btn_"):
-        return text if resolve_evdev_code(text) is not None else None
-
-    try:
-        code = int(text, 0)
-    except ValueError:
-        return None
-
-    names = evdev.ecodes.BTN.get(code)
-    if isinstance(names, (list, tuple)):
-        # evdev orders aliases oldest-first; the last name is the canonical
-        # modern label (e.g. BTN_SOUTH over BTN_A).
-        names = names[-1] if names else None
-    if isinstance(names, str) and names.startswith("BTN_"):
-        return names.lower()
-    return None
-
-
-def _resolve_gamepad_axis_target(raw: str) -> str | None:
-    """Resolve free-form axis input to a canonical evdev ABS name.
-
-    Accepts an evdev axis name (e.g. ``abs_hat0x``, ``abs_throttle``) or a
-    numeric ABS code (decimal or ``0x``-prefixed). Returns a lowercase ``abs_*``
-    name, or ``None`` when the input is not a recognized axis code.
-    """
-    text = (raw or "").strip().lower()
-    if not text:
-        return None
-
-    if text.startswith("abs_"):
-        return text if resolve_evdev_code(text) is not None else None
-
-    try:
-        code = int(text, 0)
-    except ValueError:
-        return None
-
-    names = evdev.ecodes.ABS.get(code)
-    if isinstance(names, (list, tuple)):
-        names = names[-1] if names else None
-    if isinstance(names, str) and names.startswith("ABS_"):
-        return names.lower()
-    return None
-
-
-KEYBOARD_LAYOUT = [
-    ["Esc", "F1", "F2", "F3", "F4", "F5", "F6", "F7", "F8", "F9", "F10", "F11", "F12"],
-    ["`", "1", "2", "3", "4", "5", "6", "7", "8", "9", "0", "-", "=", "Bspc"],
-    ["Tab", "Q", "W", "E", "R", "T", "Y", "U", "I", "O", "P", "[", "]", "\\"],
-    ["Caps", "A", "S", "D", "F", "G", "H", "J", "K", "L", ";", "'", "Enter"],
-    ["LShift", "Z", "X", "C", "V", "B", "N", "M", ",", ".", "/", "RShift"],
-    ["LCtrl", "LMeta", "LAlt", "Space", "RAlt", "RMeta", "Fn", "Menu", "RCtrl"],
+__all__ = [
+    "Adw",
+    "Gdk",
+    "GLib",
+    "GObject",
+    "Gtk",
+    "KeySelectorDialog",
+    "SuperkeyActionDialog",
+    "GAMEPAD_BUTTONS",
+    "KEYBOARD_LAYOUT",
+    "KEY_TO_EVDEV",
+    "KEY_WIDTHS",
+    "F_EXTRA",
+    "MPRIS_MEDIA_GROUPS",
+    "MEDIA_KEY_GROUPS",
+    "SYSTEM_KEY_GROUPS",
+    "MEDIA_KEY_TARGETS",
+    "SYSTEM_KEY_TARGETS",
+    "ACTION_DOC_LINKS",
+    "DEFAULT_RAPIDFIRE_TOOLTIP",
+    "REPEAT_CATEGORY_OPTIONS",
+    "REPEAT_RAPIDFIRE_TOOLTIP",
+    "EVDEV_TO_KEY",
+    "EVDEV_TO_GAMEPAD",
+    "_actions_docs_url",
+    "_create_actions_docs_button",
+    "_ensure_compact_tabs_css",
+    "_GamepadAxisControlsMixin",
+    "_GAMEPAD_AXIS_CUSTOM_SLOT",
+    "_PROFILE_LIFETIME_PRESETS_ENABLE",
+    "_PROFILE_LIFETIME_PRESETS_TOGGLE",
+    "_keyboard_target_allows_rapidfire",
+    "_keyboard_target_allows_tap",
+    "_resolve_gamepad_axis_target",
+    "_resolve_gamepad_button_target",
+    "session_request_async",
+    "notify_session_reload_async",
+    "get_slurp_capture",
+    "detect_compositor_sync",
+    "HardwareManager",
+    "virtual_gamepad_count",
+    "__version__",
 ]
-
-KEY_TO_EVDEV = {
-    "Esc": "key_esc",
-    "F1": "key_f1",
-    "F2": "key_f2",
-    "F3": "key_f3",
-    "F4": "key_f4",
-    "F5": "key_f5",
-    "F6": "key_f6",
-    "F7": "key_f7",
-    "F8": "key_f8",
-    "F9": "key_f9",
-    "F10": "key_f10",
-    "F11": "key_f11",
-    "F12": "key_f12",
-    "`": "key_grave",
-    "1": "key_1",
-    "2": "key_2",
-    "3": "key_3",
-    "4": "key_4",
-    "5": "key_5",
-    "6": "key_6",
-    "7": "key_7",
-    "8": "key_8",
-    "9": "key_9",
-    "0": "key_0",
-    "-": "key_minus",
-    "=": "key_equal",
-    "Bspc": "key_backspace",
-    "Tab": "key_tab",
-    "Q": "key_q",
-    "W": "key_w",
-    "E": "key_e",
-    "R": "key_r",
-    "T": "key_t",
-    "Y": "key_y",
-    "U": "key_u",
-    "I": "key_i",
-    "O": "key_o",
-    "P": "key_p",
-    "[": "key_leftbrace",
-    "]": "key_rightbrace",
-    "\\": "key_backslash",
-    "Caps": "key_capslock",
-    "A": "key_a",
-    "S": "key_s",
-    "D": "key_d",
-    "F": "key_f",
-    "G": "key_g",
-    "H": "key_h",
-    "J": "key_j",
-    "K": "key_k",
-    "L": "key_l",
-    ";": "key_semicolon",
-    "'": "key_apostrophe",
-    "Enter": "key_enter",
-    "LShift": "key_leftshift",
-    "Z": "key_z",
-    "X": "key_x",
-    "C": "key_c",
-    "V": "key_v",
-    "B": "key_b",
-    "N": "key_n",
-    "M": "key_m",
-    ",": "key_comma",
-    ".": "key_dot",
-    "/": "key_slash",
-    "RShift": "key_rightshift",
-    "LCtrl": "key_leftctrl",
-    "LMeta": "key_leftmeta",
-    "LAlt": "key_leftalt",
-    "Space": "key_space",
-    "RAlt": "key_rightalt",
-    "RMeta": "key_rightmeta",
-    "Fn": None,
-    "Menu": "key_menu",
-    "RCtrl": "key_rightctrl",
-    "F13": "key_f13",
-    "F14": "key_f14",
-    "F15": "key_f15",
-    "F16": "key_f16",
-    "F17": "key_f17",
-    "F18": "key_f18",
-    "F19": "key_f19",
-    "F20": "key_f20",
-    "F21": "key_f21",
-    "F22": "key_f22",
-    "F23": "key_f23",
-    "F24": "key_f24",
-    "Ins": "key_insert",
-    "Del": "key_delete",
-    "Home": "key_home",
-    "End": "key_end",
-    "PgUp": "key_pageup",
-    "PgDn": "key_pagedown",
-    "Up": "key_up",
-    "Down": "key_down",
-    "Left": "key_left",
-    "Right": "key_right",
-    "NumLk": "key_numlock",
-    "KP/": "key_kpslash",
-    "KP*": "key_kpasterisk",
-    "KP-": "key_kpminus",
-    "KP7": "key_kp7",
-    "KP8": "key_kp8",
-    "KP9": "key_kp9",
-    "KP+": "key_kpplus",
-    "KP4": "key_kp4",
-    "KP5": "key_kp5",
-    "KP6": "key_kp6",
-    "KP1": "key_kp1",
-    "KP2": "key_kp2",
-    "KP3": "key_kp3",
-    "KPEnter": "key_kpenter",
-    "KP0": "key_kp0",
-    "KP.": "key_kpdot",
-    "Mute": "key_mute",
-    "Volume Down": "key_volumedown",
-    "Volume Up": "key_volumeup",
-    "Mic Mute": "key_micmute",
-    "Play/Pause": "key_playpause",
-    "Play": "key_play",
-    "Pause": "key_pause",
-    "Stop": "key_stop",
-    "Previous Track": "key_previoussong",
-    "Next Track": "key_nextsong",
-}
-
-KEY_WIDTHS = {
-    "Esc": 1,
-    "Bspc": 2,
-    "Tab": 1.5,
-    "\\": 1.5,
-    "Caps": 1.75,
-    "Enter": 2.25,
-    "LShift": 2.25,
-    "RShift": 2.75,
-    "LCtrl": 1.25,
-    "LMeta": 1.25,
-    "LAlt": 1.25,
-    "Space": 6.25,
-    "RAlt": 1.25,
-    "RMeta": 1.25,
-    "Fn": 1.25,
-    "Menu": 1.25,
-    "RCtrl": 1.25,
-}
-
-F_EXTRA = ["F13", "F14", "F15", "F16", "F17", "F18", "F19", "F20", "F21", "F22", "F23", "F24"]
-
-MPRIS_MEDIA_GROUPS = [
-    (
-        "Player Controls",
-        [
-            ("Previous", "previous", "media-skip-backward-symbolic"),
-            ("Play/Pause", "play_pause", "media-playback-start-symbolic"),
-            ("Next", "next", "media-skip-forward-symbolic"),
-            ("Stop", "stop", "media-playback-stop-symbolic"),
-            ("Play", "play", "media-playback-start-symbolic"),
-            ("Pause", "pause", "media-playback-pause-symbolic"),
-        ],
-    )
-]
-
-MEDIA_KEY_GROUPS = [
-    (
-        "Playback",
-        [
-            ("Previous", "key_previoussong", "media-skip-backward-symbolic"),
-            ("Play/Pause", "key_playpause", "media-playback-start-symbolic"),
-            ("Next", "key_nextsong", "media-skip-forward-symbolic"),
-            ("Stop", "key_stop", "media-playback-stop-symbolic"),
-            ("Play", "key_play", "media-playback-start-symbolic"),
-            ("Pause", "key_pause", "media-playback-pause-symbolic"),
-        ],
-    ),
-]
-SYSTEM_KEY_GROUPS = [
-    (
-        "System Keys",
-        [
-            ("Vol Down", "key_volumedown", "audio-volume-low-symbolic"),
-            ("Vol Up", "key_volumeup", "audio-volume-high-symbolic"),
-            ("Mute", "key_mute", "audio-volume-muted-symbolic"),
-            ("Mic Mute", "key_micmute", "microphone-sensitivity-muted-symbolic"),
-            ("Bright Down", "key_brightnessdown", "display-brightness-symbolic"),
-            ("Bright Up", "key_brightnessup", "display-brightness-symbolic"),
-        ],
-    )
-]
-MEDIA_KEY_TARGETS = {
-    evdev_id for _title, buttons in MEDIA_KEY_GROUPS for _label, evdev_id, _icon_name in buttons
-}
-SYSTEM_KEY_TARGETS = {
-    evdev_id for _title, buttons in SYSTEM_KEY_GROUPS for _label, evdev_id, _icon_name in buttons
-}
-
-
-def _keyboard_target_allows_rapidfire(evdev_name: str) -> bool:
-    return evdev_name not in MEDIA_KEY_TARGETS
-
-
-def _keyboard_target_allows_tap(evdev_name: str) -> bool:
-    return evdev_name not in MEDIA_KEY_TARGETS
-
-ACTION_DOC_LINKS = {
-    "analog_presets": ("analog-controls", "Analog Controls"),
-    "analog_control": ("analog-controls", "Analog Controls"),
-    "special": ("special", "Special"),
-    "keyboard": ("keyboard", "Keyboard"),
-    "navigation": ("navigation", "Navigation"),
-    "media": ("media", "Media"),
-    "mouse": ("mouse", "Mouse"),
-    "gamepad": ("gamepad", "Gamepad"),
-    "hyprland": ("hyprland", "Hyprland"),
-    "niri": ("niri", "Niri"),
-    "kde": ("kde-plasma", "KDE Plasma"),
-    "gnome": ("gnome", "GNOME"),
-    "superkey": ("super-keys", "Super Keys"),
-    "macro": ("macro", "Macro"),
-    "profile": ("profile", "Profile"),
-    "exec": ("execute-shell-command", "Command"),
-}
-
-REPEAT_RAPIDFIRE_TOOLTIP = (
-    "Rapidfire repeats only remembered keyboard keys, mouse buttons, mouse wheel actions, "
-    "and gamepad buttons. Other remembered actions run once."
-)
-DEFAULT_RAPIDFIRE_TOOLTIP = "Repeatedly send the mapped action while the button is held"
-REPEAT_CATEGORY_OPTIONS = (
-    (REPEAT_CATEGORY_KEYBOARD, "Keys", "Repeat remembered keyboard actions"),
-    (REPEAT_CATEGORY_MOUSE, "Mouse", "Repeat remembered mouse button and wheel actions"),
-    (REPEAT_CATEGORY_GAMEPAD, "Gamepad", "Repeat remembered gamepad actions"),
-    (REPEAT_CATEGORY_MACRO, "Macros", "Repeat remembered macro actions"),
-    (
-        REPEAT_CATEGORY_SPECIAL,
-        "Other",
-        "Repeat remembered Keymasq special actions that do not fit the other groups, "
-        "including mouse movement, shell commands, and recording or playback controls.",
-    ),
-)
-
-EVDEV_TO_KEY = {v: k for k, v in KEY_TO_EVDEV.items()}
-EVDEV_TO_GAMEPAD = {v: k for k, v in GAMEPAD_BUTTONS.items()}
-
-_compact_tabs_css_installed = False
-
-
-def _actions_docs_url(anchor: str) -> str:
-    return actions_docs_url(anchor, version=__version__)
-
-
-def _create_actions_docs_button() -> Gtk.Button:
-    btn = Gtk.Button(label="?")
-    btn.add_css_class("flat")
-    btn.add_css_class("actions-docs-button")
-    btn.set_tooltip_text("Open documentation for this tab")
-    return btn
-
-
-def _ensure_compact_tabs_css() -> None:
-    global _compact_tabs_css_installed
-    if _compact_tabs_css_installed:
-        return
-
-    display = Gdk.Display.get_default()
-    if display is None:
-        return
-
-    provider = Gtk.CssProvider()
-    provider.load_from_string(
-        """
-        .compact-map-tabs button {
-            padding-left: 7px;
-            padding-right: 7px;
-            min-height: 28px;
-            font-size: 0.85em;
-        }
-        """
-    )
-    Gtk.StyleContext.add_provider_for_display(
-        display,
-        provider,
-        Gtk.STYLE_PROVIDER_PRIORITY_APPLICATION,
-    )
-    _compact_tabs_css_installed = True
-
-
-_GAMEPAD_AXIS_CUSTOM_SLOT = "custom"
-
-
-class _GamepadAxisControlsMixin:
-    gamepad_axis_targets: list[str]
-    gamepad_axis_dropdown: Gtk.DropDown
-    gamepad_axis_custom_entry: Gtk.Entry
-    gamepad_axis_value: Gtk.SpinButton
-    gamepad_axis_percent: Gtk.SpinButton
-    gamepad_axis_percent_label: Gtk.Label
-    gamepad_code_entry: Gtk.Entry
-    _syncing_gamepad_axis_controls: bool
-
-    def _build_gamepad_axis_controls(self) -> Gtk.Widget:
-        row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        row.set_halign(Gtk.Align.CENTER)
-
-        self.gamepad_axis_targets = [*GAMEPAD_AXIS_RANGES, _GAMEPAD_AXIS_CUSTOM_SLOT]
-        labels = [GAMEPAD_AXIS_RANGES[target].label for target in GAMEPAD_AXIS_RANGES]
-        labels.append("Custom")
-        self.gamepad_axis_dropdown = Gtk.DropDown.new_from_strings(labels)
-        self.gamepad_axis_dropdown.connect("notify::selected", self._on_gamepad_axis_changed)
-        row.append(self.gamepad_axis_dropdown)
-
-        self.gamepad_axis_custom_entry = Gtk.Entry()
-        self.gamepad_axis_custom_entry.set_placeholder_text("abs_ code")
-        self.gamepad_axis_custom_entry.set_width_chars(11)
-        self.gamepad_axis_custom_entry.set_tooltip_text(
-            "Custom axis evdev name (e.g. abs_hat0x, abs_hat0y, abs_throttle, "
-            "abs_rudder) or numeric code."
-        )
-        row.append(self.gamepad_axis_custom_entry)
-
-        self.gamepad_axis_value = Gtk.SpinButton()
-        self.gamepad_axis_value.set_numeric(True)
-        self.gamepad_axis_value.set_increments(1, 256)
-        self.gamepad_axis_value.set_tooltip_text("Raw axis value")
-        self.gamepad_axis_value.connect("value-changed", self._on_gamepad_axis_value_changed)
-        row.append(self.gamepad_axis_value)
-
-        self.gamepad_axis_percent = Gtk.SpinButton(
-            adjustment=Gtk.Adjustment(value=100, lower=-100, upper=100, step_increment=1)
-        )
-        self.gamepad_axis_percent.set_numeric(True)
-        self.gamepad_axis_percent.set_tooltip_text("Percent")
-        self.gamepad_axis_percent.connect("value-changed", self._on_gamepad_axis_percent_changed)
-        row.append(self.gamepad_axis_percent)
-        self.gamepad_axis_percent_label = Gtk.Label(label="%")
-        row.append(self.gamepad_axis_percent_label)
-
-        apply_btn = Gtk.Button(label="Map Analog")
-        apply_btn.add_css_class("suggested-action")
-        apply_btn.connect("clicked", self._on_gamepad_axis_apply_clicked)
-        row.append(apply_btn)
-
-        self._on_gamepad_axis_changed(self.gamepad_axis_dropdown, None)
-        return row
-
-    def _selected_gamepad_axis_slot(self) -> str:
-        index = int(self.gamepad_axis_dropdown.get_selected())
-        if index < 0 or index >= len(self.gamepad_axis_targets):
-            return "abs_x"
-        return self.gamepad_axis_targets[index]
-
-    def _selected_gamepad_axis_target(self) -> str | None:
-        slot = self._selected_gamepad_axis_slot()
-        if slot == _GAMEPAD_AXIS_CUSTOM_SLOT:
-            return _resolve_gamepad_axis_target(self.gamepad_axis_custom_entry.get_text())
-        return slot
-
-    def _on_gamepad_axis_changed(self, dropdown, _param) -> None:
-        is_custom = self._selected_gamepad_axis_slot() == _GAMEPAD_AXIS_CUSTOM_SLOT
-        self.gamepad_axis_custom_entry.set_visible(is_custom)
-        self.gamepad_axis_percent.set_visible(not is_custom)
-        self.gamepad_axis_percent_label.set_visible(not is_custom)
-
-        if is_custom:
-            current = int(self.gamepad_axis_value.get_value())
-            self._syncing_gamepad_axis_controls = True
-            self.gamepad_axis_value.set_adjustment(
-                Gtk.Adjustment(
-                    value=current,
-                    lower=-2147483648,
-                    upper=2147483647,
-                    step_increment=1,
-                    page_increment=256,
-                )
-            )
-            self._syncing_gamepad_axis_controls = False
-            return
-
-        axis = gamepad_axis_range(self._selected_gamepad_axis_slot())
-        if axis is None:
-            return
-        self._syncing_gamepad_axis_controls = True
-        self.gamepad_axis_value.set_adjustment(
-            Gtk.Adjustment(
-                value=axis.maximum,
-                lower=axis.minimum,
-                upper=axis.maximum,
-                step_increment=1,
-                page_increment=256,
-            )
-        )
-        lower = -100 if axis.minimum < 0 else 0
-        self.gamepad_axis_percent.set_adjustment(
-            Gtk.Adjustment(value=100, lower=lower, upper=100, step_increment=1)
-        )
-        self._syncing_gamepad_axis_controls = False
-
-    def _on_gamepad_axis_percent_changed(self, spin: Gtk.SpinButton) -> None:
-        if getattr(self, "_syncing_gamepad_axis_controls", False):
-            return
-        target = self._selected_gamepad_axis_target()
-        if target is None:
-            return
-        self._syncing_gamepad_axis_controls = True
-        self.gamepad_axis_value.set_value(
-            gamepad_axis_value_from_percent(target, spin.get_value())
-        )
-        self._syncing_gamepad_axis_controls = False
-
-    def _on_gamepad_axis_value_changed(self, spin: Gtk.SpinButton) -> None:
-        if getattr(self, "_syncing_gamepad_axis_controls", False):
-            return
-        if self._selected_gamepad_axis_slot() == _GAMEPAD_AXIS_CUSTOM_SLOT:
-            return
-        target = self._selected_gamepad_axis_target()
-        if target is None:
-            return
-        self._syncing_gamepad_axis_controls = True
-        self.gamepad_axis_percent.set_value(
-            gamepad_axis_percent_from_value(target, spin.get_value())
-        )
-        self._syncing_gamepad_axis_controls = False
-
-    def _on_gamepad_axis_apply_clicked(self, btn) -> None:
-        target = self._selected_gamepad_axis_target()
-        if not target:
-            self.gamepad_axis_custom_entry.set_text("")
-            self.gamepad_axis_custom_entry.set_placeholder_text("Unknown axis code")
-            return
-        self._on_gamepad_axis_clicked(
-            btn,
-            target,
-            int(self.gamepad_axis_value.get_value()),
-        )
-
-    def _prefill_gamepad_inputs(self) -> None:
-        """Restore the code/axis fields when editing an existing gamepad action."""
-        action = getattr(self, "_current_action", None)
-        if action is None:
-            return
-        action_type = getattr(action, "action_type", None)
-        target = str(getattr(action, "target", "") or "")
-
-        if action_type == ActionType.GAMEPAD:
-            # Pre-fill the free-form code field for buttons outside the template.
-            if target and target not in EVDEV_TO_GAMEPAD and hasattr(self, "gamepad_code_entry"):
-                self.gamepad_code_entry.set_text(target)
-            return
-
-        if action_type == ActionType.GAMEPAD_AXIS:
-            self._prefill_gamepad_axis(target, int(getattr(action, "axis_value", 0) or 0))
-
-    def _prefill_gamepad_axis(self, target: str, value: int) -> None:
-        normalized = normalize_gamepad_axis_target(target) or target
-        if normalized in GAMEPAD_AXIS_RANGES:
-            self.gamepad_axis_dropdown.set_selected(
-                self.gamepad_axis_targets.index(normalized)
-            )
-        elif _resolve_gamepad_axis_target(normalized):
-            self.gamepad_axis_dropdown.set_selected(
-                self.gamepad_axis_targets.index(_GAMEPAD_AXIS_CUSTOM_SLOT)
-            )
-            self.gamepad_axis_custom_entry.set_text(normalized)
-        else:
-            return
-        self.gamepad_axis_value.set_value(value)
-
-    def _on_gamepad_axis_clicked(self, btn, axis_target: str, axis_value: int) -> None:
-        raise NotImplementedError
-
-
-class KeySelectorDialog(Adw.Dialog, _GamepadAxisControlsMixin):
-    __gsignals__ = {
-        "key-selected": (GObject.SignalFlags.RUN_FIRST, None, (object,)),
-    }
-
-    def __init__(
-        self,
-        parent: Gtk.Widget,
-        button_label: str,
-        current_action: MappingAction | None = None,
-        compositor_action_status: dict[str, object] | None = None,
-        *,
-        allow_passthrough: bool = True,
-        allow_clear_mapping: bool = True,
-        allow_suppress: bool = True,
-        allow_superkey: bool = True,
-        allow_repeat: bool = True,
-        allow_rapidfire: bool = True,
-        allow_tap: bool = True,
-        allow_macro_options: bool = True,
-        source_type: str = "button",
-        analog_input_type: str | None = None,
-    ):
-        super().__init__(title=f"Map: {button_label}", content_width=570, content_height=580)
-        self._parent = parent
-        self._button_label = button_label
-        self._current_action = current_action
-        self._allow_passthrough = allow_passthrough
-        self._allow_clear_mapping = allow_clear_mapping
-        self._allow_suppress = allow_suppress
-        self._allow_superkey = allow_superkey
-        self._allow_repeat = allow_repeat
-        self._allow_rapidfire = allow_rapidfire
-        self._allow_tap = allow_tap
-        self._allow_macro_options = allow_macro_options
-        self._source_type = str(source_type or "button")
-        self._analog_input_type = (
-            str(analog_input_type or "").lower()
-            if self._source_type == "analog" and analog_input_type
-            else None
-        )
-        self._compositor_action_status = self._resolve_compositor_action_status(
-            compositor_action_status
-        )
-        self._rapidfire_enabled = False
-        self._rapidfire_hold = 20
-        self._rapidfire_wait = 20
-        self._tap_enabled = False
-        self._tap_hold = 50
-        self._repeat_categories: list[str] = list(DEFAULT_REPEAT_CATEGORIES)
-        self._repeat_toggle_buttons: dict[str, Gtk.ToggleButton] = {}
-        self._repeat_button: Gtk.ToggleButton | None = None
-        self._repeat_map_btn: Gtk.Button | None = None
-        self._repeat_options_box: Gtk.Widget | None = None
-        self._macro_list: list[dict] = []
-        self._selected_macro: str | None = None
-        self._cancel_macro_playback_btn: Gtk.Button | None = None
-        self._macro_recording_enabled = self._resolve_macro_recording_enabled(default=False)
-        self._macro_slot_console: Gtk.Box | None = None
-        self._superkey_list: list[SuperkeyConfig] = []
-        self._superkey_names: list[str] = []
-        self._selected_superkey: str | None = None
-        self._analog_control_list: list[AnalogControlConfig] = []
-        self._analog_control_names: list[str] = []
-        self._selected_analog_control: str | None = None
-        self._selected_analog_controls: list[str] = []
-        self._macro_replay_movement: bool = True
-        self._macro_replay_clicks: bool = True
-        self._macro_speed: float = 1.0
-        self._profile_entries: list[dict] = []
-        self._selected_profile_action: str = "enable"
-        self._selected_profile_name: str = ""
-        self._profile_lifetime_preset: str = "until_changed"
-        self._profile_lifetime_action_count: int = 2
-        self._profile_lifetime_timeout_ms: int = 1500
-        self._profile_custom_trigger_end: bool = False
-        self._profile_custom_action_count: bool = False
-        self._profile_custom_timeout: bool = False
-        self._profile_lifetime_selection_updating: bool = False
-        self._profile_lifetime_model_keys: list[str] = []
-        self._selected_gamepad_output_id: str | None = (
-            current_action.output_id
-            if current_action
-            and current_action.action_type in (ActionType.GAMEPAD, ActionType.GAMEPAD_AXIS)
-            else None
-        )
-        self._gamepad_output_ids: list[str | None] = []
-        self._gamepad_output_dropdown: Gtk.DropDown | None = None
-        self._gamepad_output_header: Gtk.Widget | None = None
-        self._gamepad_output_warning_label: Gtk.Label | None = None
-        self._profile_name_items: list[str] = []
-        self._profile_name_populating: bool = False
-        self._exec_cmd: str = ""
-        self._mouse_move_x: int = 0
-        self._mouse_move_y: int = 0
-        self._mouse_move_mode: str = "rel"
-        self._capture_delay_seconds: float = 2.0
-        self._capture_timeout_id: int = 0
-        self._capture_pending: bool = False
-        self._capture_request_id: int = 0
-        self._capture_apply: Callable[[int, int], None] | None = None
-        self._capture_status_label: Gtk.Label | None = None
-        self._capture_button: Gtk.Button | None = None
-        self._slurp_capture = get_slurp_capture()
-        self._slurp_capture.set_compositor(detect_compositor_sync())
-        self._slurp_available = self._slurp_capture.available
-
-        if current_action:
-            self._rapidfire_enabled = current_action.rapidfire_enabled
-            self._rapidfire_hold = current_action.rapidfire_hold_ms
-            self._rapidfire_wait = current_action.rapidfire_wait_ms
-            self._tap_enabled = current_action.tap_enabled
-            self._tap_hold = current_action.tap_hold_ms
-            if current_action.action_type == ActionType.MACRO:
-                self._selected_macro = current_action.macro_name
-                self._macro_replay_movement = current_action.macro_replay_mouse_movement
-                self._macro_replay_clicks = current_action.macro_replay_mouse_clicks
-                self._macro_speed = current_action.macro_speed
-            elif current_action.action_type == ActionType.SUPERKEY:
-                self._selected_superkey = current_action.superkey_name
-            elif current_action.action_type == ActionType.ANALOG_CONTROL:
-                self._selected_analog_controls = list(current_action.analog_control_names)
-                if not self._selected_analog_controls and current_action.analog_control_name:
-                    self._selected_analog_controls = [current_action.analog_control_name]
-                self._selected_analog_control = (
-                    self._selected_analog_controls[0]
-                    if self._selected_analog_controls
-                    else None
-                )
-            elif current_action.action_type == ActionType.EXEC:
-                self._exec_cmd = current_action.cmd or ""
-            elif current_action.action_type == ActionType.REPEAT:
-                self._repeat_categories = list(
-                    current_action.repeat_categories
-                    if current_action.repeat_categories is not None
-                    else DEFAULT_REPEAT_CATEGORIES
-                )
-            elif current_action.action_type in (
-                ActionType.PROFILE_ENABLE,
-                ActionType.PROFILE_DISABLE,
-                ActionType.PROFILE_TOGGLE,
-            ):
-                self._selected_profile_name = str(
-                    current_action.profile_name or current_action.target or ""
-                )
-                self._restore_profile_lifetime(current_action.profile_deactivation)
-                if current_action.action_type == ActionType.PROFILE_ENABLE:
-                    self._selected_profile_action = "enable"
-                elif current_action.action_type == ActionType.PROFILE_DISABLE:
-                    self._selected_profile_action = "disable"
-                else:
-                    self._selected_profile_action = "toggle"
-            elif current_action.action_type in (
-                ActionType.MOUSE_MOVE_REL,
-                ActionType.MOUSE_MOVE_ABS,
-            ):
-                self._mouse_move_x = int(current_action.move_x)
-                self._mouse_move_y = int(current_action.move_y)
-                if current_action.action_type == ActionType.MOUSE_MOVE_ABS:
-                    self._mouse_move_mode = "abs"
-        if not self._allow_rapidfire:
-            self._rapidfire_enabled = False
-        if not self._allow_tap:
-            self._tap_enabled = False
-
-        self._build_ui()
-
-    def _restore_profile_lifetime(
-        self,
-        policy: ProfileDeactivationPolicy | None,
-    ) -> None:
-        if policy is None:
-            self._profile_lifetime_preset = "until_changed"
-            return
-
-        self._profile_lifetime_action_count = int(policy.after_actions or 2)
-        self._profile_lifetime_timeout_ms = int(policy.timeout_ms or 1500)
-        self._profile_custom_trigger_end = bool(policy.on_trigger_end)
-        self._profile_custom_action_count = policy.after_actions is not None
-        self._profile_custom_timeout = policy.timeout_ms is not None
-
-        simple_trigger = (
-            policy.on_trigger_end
-            and policy.after_actions is None
-            and policy.timeout_ms is None
-        )
-        one_shot = (
-            not policy.on_trigger_end
-            and policy.after_actions == 1
-        )
-        simple_count = (
-            not policy.on_trigger_end
-            and policy.after_actions is not None
-            and policy.timeout_ms is None
-        )
-        if simple_trigger:
-            self._profile_lifetime_preset = "while_trigger_active"
-        elif one_shot:
-            self._profile_lifetime_preset = "after_one_action"
-        elif simple_count:
-            self._profile_lifetime_preset = "custom"
-        else:
-            self._profile_lifetime_preset = "custom"
-
-    def _build_ui(self):
-        _ensure_compact_tabs_css()
-
-        main_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=0)
-        main_box.set_margin_top(12)
-        main_box.set_margin_bottom(12)
-        main_box.set_margin_start(12)
-        main_box.set_margin_end(12)
-
-        self.options_box = self._build_options_box()
-        self._compositor_action_pages = build_compositor_action_pages(
-            self._current_action,
-            self._on_compositor_action_selected,
-            self._compositor_action_status,
-            capture_position=self._capture_compositor_position,
-        )
-        self._compositor_action_page_ids = {page.page_id for page in self._compositor_action_pages}
-
-        self.stack = Gtk.Stack()
-        self.stack.set_vexpand(True)
-        if self._source_type == "analog":
-            self.stack.add_titled(
-                self._build_analog_presets_tab(),
-                "analog_presets",
-                "Presets",
-            )
-            self.stack.add_titled(
-                self._build_analog_control_tab(),
-                "analog_control",
-                "Analog Controls",
-            )
-            self.stack.add_titled(self._build_special_tab(), "special", "Special")
-        else:
-            self.stack.add_titled(self._build_special_tab(), "special", "Special")
-            self.stack.add_titled(self._build_keyboard_tab(), "keyboard", "Keyboard")
-            self.stack.add_titled(self._build_navigation_tab(), "navigation", "Navigation")
-            self.stack.add_titled(self._build_media_tab(), "media", "Media")
-            self.stack.add_titled(self._build_mouse_tab(), "mouse", "Mouse")
-            for page in self._compositor_action_pages:
-                self.stack.add_titled(page.widget, page.page_id, page.title)
-            self.stack.add_titled(self._build_gamepad_tab(), "gamepad", "Gamepad")
-            if self._allow_superkey:
-                self.stack.add_titled(self._build_superkey_tab(), "superkey", "Super Keys")
-            self.stack.add_titled(self._build_macro_tab(), "macro", "Macro")
-            self.stack.add_titled(self._build_profile_tab(), "profile", "Profile")
-
-        self._set_initial_tab()
-
-        frame = Gtk.Frame()
-        frame.set_vexpand(True)
-
-        inner = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=0)
-
-        title_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        title_row.set_halign(Gtk.Align.CENTER)
-        title_row.set_margin_top(12)
-        title_row.set_margin_bottom(6)
-
-        title_label = Gtk.Label(label=f"Map: {self._button_label}")
-        title_label.add_css_class("title-3")
-        title_label.set_halign(Gtk.Align.CENTER)
-        title_row.append(title_label)
-
-        self._gamepad_output_header = self._build_gamepad_output_header()
-        if self._gamepad_output_header is not None:
-            title_row.append(self._gamepad_output_header)
-
-        inner.append(title_row)
-
-        if self._current_action:
-            current_label = Gtk.Label(label=self._describe_current_action())
-            current_label.add_css_class("dim-label")
-            current_label.set_halign(Gtk.Align.CENTER)
-            current_label.set_margin_bottom(10)
-            inner.append(current_label)
-        else:
-            title_row.set_margin_bottom(12)
-
-        inner.append(Gtk.Separator())
-
-        sidebar = Gtk.StackSidebar()
-        sidebar.set_stack(self.stack)
-        sidebar.set_size_request(120, -1)
-
-        paned = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=0)
-        paned.set_vexpand(True)
-        paned.append(sidebar)
-        paned.append(Gtk.Separator(orientation=Gtk.Orientation.VERTICAL))
-
-        right_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=0)
-        right_box.set_hexpand(True)
-        right_box.append(self.stack)
-        right_box.append(Gtk.Separator())
-
-        options_pad = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=0)
-        options_pad.set_margin_top(8)
-        options_pad.set_margin_bottom(8)
-        options_pad.set_margin_start(12)
-        options_pad.set_margin_end(12)
-        options_pad.append(self.options_box)
-        right_box.append(options_pad)
-
-        paned.append(right_box)
-        inner.append(paned)
-
-        inner.append(Gtk.Separator())
-
-        footer = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        footer.set_hexpand(True)
-        footer.set_margin_top(8)
-        footer.set_margin_bottom(8)
-        footer.set_margin_start(12)
-        footer.set_margin_end(12)
-
-        self.actions_docs_btn = _create_actions_docs_button()
-        self.actions_docs_btn.connect("clicked", self._on_actions_docs_clicked)
-        footer.append(self.actions_docs_btn)
-
-        footer_spacer = Gtk.Box()
-        footer_spacer.set_hexpand(True)
-        footer.append(footer_spacer)
-
-        cancel_macro_playback_btn = self._create_cancel_macro_playback_button()
-        cancel_macro_playback_btn.set_visible(False)
-        self._cancel_macro_playback_btn = cancel_macro_playback_btn
-        footer.append(cancel_macro_playback_btn)
-
-        self.map_btn = Gtk.Button(label="Map")
-        self.map_btn.add_css_class("suggested-action")
-        self.map_btn.set_sensitive(False)
-        self.map_btn.set_visible(False)
-        self.map_btn.connect("clicked", self._on_map_clicked)
-        footer.append(self.map_btn)
-
-        cancel_btn = Gtk.Button(label="Cancel")
-        cancel_btn.connect("clicked", self._on_cancel_clicked)
-        footer.append(cancel_btn)
-        inner.append(footer)
-
-        frame.set_child(inner)
-        main_box.append(frame)
-        self.set_child(main_box)
-
-        self.stack.connect("notify::visible-child", self._on_tab_changed)
-        self._on_tab_changed(self.stack, None)
-
-    def _on_cancel_clicked(self, _button: Gtk.Button) -> None:
-        self.close()
-
-    def _create_cancel_macro_playback_button(self) -> Gtk.Button:
-        content = Adw.ButtonContent(
-            icon_name="media-playback-stop-symbolic",
-            label="Cancel Macro Playback",
-        )
-        content.add_css_class("macro-stop-icon")
-        button = Gtk.Button()
-        button.set_child(content)
-        button.connect(
-            "clicked",
-            self._on_macro_special_action_clicked,
-            "cancel_macro_playback",
-        )
-        button.set_tooltip_text("Stop currently running macro playback")
-        return button
-
-    def _macro_parent_candidates(self) -> list[object]:
-        candidates: list[object] = []
-        queue: list[object] = [self._parent]
-        seen: set[int] = set()
-        while queue:
-            candidate = queue.pop(0)
-            if candidate is None:
-                continue
-            candidate_id = id(candidate)
-            if candidate_id in seen:
-                continue
-            seen.add(candidate_id)
-            candidates.append(candidate)
-
-            main_window = getattr(candidate, "main_window", None)
-            if main_window is not None:
-                queue.append(main_window)
-            parent = getattr(candidate, "_parent", None)
-            if parent is not None:
-                queue.append(parent)
-            get_root = getattr(candidate, "get_root", None)
-            if callable(get_root):
-                root = get_root()
-                if root is not None:
-                    queue.append(root)
-        return candidates
-
-    def _profile_manager_for_child_dialog(self) -> ProfileManager | None:
-        for candidate in self._macro_parent_candidates():
-            profile_manager = getattr(candidate, "profile_manager", None)
-            if profile_manager is not None:
-                return cast(ProfileManager, profile_manager)
-        return None
-
-    def _resolve_macro_recording_enabled(self, *, default: bool) -> bool:
-        for candidate in self._macro_parent_candidates():
-            enabled = getattr(candidate, "macro_recording_enabled", None)
-            if callable(enabled):
-                return bool(enabled())
-            raw_enabled = getattr(candidate, "_macro_recording_enabled", None)
-            if isinstance(raw_enabled, bool):
-                return raw_enabled
-        return default
-
-    def _present_macro_recording_settings(self, _button: Gtk.Button) -> None:
-        for candidate in self._macro_parent_candidates():
-            present_settings = getattr(candidate, "present_recording_settings_dialog", None)
-            if callable(present_settings):
-                present_settings(reason="settings")
-                self.close()
-                return
-
-    def _build_special_tab(self) -> Gtk.Widget:
-        box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=16)
-        box.set_margin_top(16)
-        box.set_margin_bottom(16)
-        box.set_margin_start(16)
-        box.set_margin_end(16)
-        box.set_valign(Gtk.Align.CENTER)
-
-        special_buttons_added = False
-
-        if self._allow_clear_mapping:
-            clear_label = "Clear Mapping" if self._source_type == "analog" else "Passthrough"
-            clear_btn = self._create_key_button(clear_label, "clear_mapping", large=True)
-            clear_btn.connect("clicked", self._on_special_clicked, "clear_mapping")
-            clear_btn.set_tooltip_text(
-                "Do not store a mapping here, so lower-priority profiles can still apply"
-            )
-            box.append(clear_btn)
-            special_buttons_added = True
-
-        if self._allow_suppress:
-            suppress_btn = self._create_key_button("Suppress", "suppress", large=True)
-            suppress_btn.connect("clicked", self._on_special_clicked, "suppress")
-            suppress_btn.set_tooltip_text("Block the button press entirely — nothing is sent")
-            box.append(suppress_btn)
-            special_buttons_added = True
-
-        if self._source_type == "analog":
-            return box
-
-        if self._allow_repeat:
-            repeat_btn = Gtk.ToggleButton(label="Repeat Last Action")
-            repeat_btn.add_css_class("key-button")
-            repeat_btn.set_size_request(200, 50)
-            repeat_btn.set_active(
-                bool(
-                    self._current_action
-                    and self._current_action.action_type == ActionType.REPEAT
-                )
-            )
-            repeat_btn.connect("toggled", self._on_repeat_button_toggled)
-            repeat_btn.set_tooltip_text("Replay the last remembered mapped action")
-            self._repeat_button = repeat_btn
-            box.append(repeat_btn)
-            box.append(self._build_repeat_section())
-            special_buttons_added = True
-
-        exec_label = Gtk.Label(label="Execute Shell Command")
-        exec_label.add_css_class("dim-label")
-        if special_buttons_added:
-            box.append(Gtk.Separator())
-        box.append(exec_label)
-
-        exec_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        exec_box.set_halign(Gtk.Align.CENTER)
-
-        self.exec_entry = Gtk.Entry()
-        self.exec_entry.set_placeholder_text("e.g., notify-send 'hello'")
-        self.exec_entry.set_size_request(300, -1)
-        self.exec_entry.set_text(self._exec_cmd)
-        self.exec_entry.connect("changed", self._on_exec_text_changed)
-        exec_box.append(self.exec_entry)
-
-        self.exec_map_btn = Gtk.Button(label="Map Command")
-        self.exec_map_btn.add_css_class("suggested-action")
-        self.exec_map_btn.set_sensitive(bool(self._exec_cmd.strip()))
-        self.exec_map_btn.connect("clicked", self._on_exec_map_clicked)
-        exec_box.append(self.exec_map_btn)
-
-        box.append(exec_box)
-
-        return box
-
-    def _build_repeat_section(self) -> Gtk.Widget:
-        section = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=8)
-        section.set_halign(Gtk.Align.CENTER)
-
-        self._repeat_options_box = section
-        section.set_visible(bool(self._repeat_button and self._repeat_button.get_active()))
-
-        section.append(self._build_repeat_rapidfire_row())
-
-        self._repeat_toggle_buttons = {}
-        categories_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL)
-        categories_box.add_css_class("linked")
-        categories_box.set_halign(Gtk.Align.CENTER)
-        categories_box.set_hexpand(True)
-        categories_box.set_homogeneous(True)
-        categories_box.set_size_request(355, -1)
-        categories_box.set_margin_top(4)
-        for category, label, tooltip in REPEAT_CATEGORY_OPTIONS:
-            toggle = Gtk.ToggleButton(label=label)
-            toggle.set_active(category in self._repeat_categories)
-            toggle.set_tooltip_text(tooltip)
-            toggle.set_hexpand(True)
-            toggle.connect("toggled", self._on_repeat_category_toggled, category)
-            categories_box.append(toggle)
-            self._repeat_toggle_buttons[category] = toggle
-
-        section.append(categories_box)
-
-        action_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        action_row.set_halign(Gtk.Align.CENTER)
-        action_row.set_margin_top(4)
-        repeat_map_btn = Gtk.Button(label="Map Repeat")
-        repeat_map_btn.add_css_class("suggested-action")
-        repeat_map_btn.connect("clicked", self._on_repeat_map_clicked)
-        self._repeat_map_btn = repeat_map_btn
-        action_row.append(repeat_map_btn)
-        section.append(action_row)
-        self._update_repeat_map_button()
-
-        return section
-
-    def _build_repeat_rapidfire_row(self) -> Gtk.Widget:
-        row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        row.set_halign(Gtk.Align.CENTER)
-
-        self.repeat_rapidfire_check = Gtk.CheckButton(label="Rapidfire")
-        self.repeat_rapidfire_check.set_active(self._allow_rapidfire and self._rapidfire_enabled)
-        self.repeat_rapidfire_check.set_tooltip_text(REPEAT_RAPIDFIRE_TOOLTIP)
-        self.repeat_rapidfire_check.connect("toggled", self._on_repeat_rapidfire_toggled)
-        row.append(self.repeat_rapidfire_check)
-
-        self.repeat_hold_label = Gtk.Label(label="Hold:")
-        row.append(self.repeat_hold_label)
-        self.repeat_hold_spin = Gtk.SpinButton()
-        self.repeat_hold_spin.set_adjustment(
-            Gtk.Adjustment(
-                value=self._rapidfire_hold,
-                lower=MIN_RAPIDFIRE_HOLD_MS,
-                upper=1000,
-                step_increment=1,
-            )
-        )
-        row.append(self.repeat_hold_spin)
-        self.repeat_hold_ms_label = Gtk.Label(label="ms")
-        row.append(self.repeat_hold_ms_label)
-
-        self.repeat_wait_label = Gtk.Label(label="Wait:")
-        row.append(self.repeat_wait_label)
-        self.repeat_wait_spin = Gtk.SpinButton()
-        self.repeat_wait_spin.set_adjustment(
-            Gtk.Adjustment(
-                value=self._rapidfire_wait,
-                lower=MIN_RAPIDFIRE_WAIT_MS,
-                upper=1000,
-                step_increment=1,
-            )
-        )
-        row.append(self.repeat_wait_spin)
-        self.repeat_wait_ms_label = Gtk.Label(label="ms")
-        row.append(self.repeat_wait_ms_label)
-
-        self._update_repeat_rapidfire_visibility()
-        return row
-
-    def _update_repeat_rapidfire_visibility(self) -> None:
-        self.repeat_rapidfire_check.set_visible(self._allow_rapidfire)
-        rf_active = self._allow_rapidfire and self.repeat_rapidfire_check.get_active()
-        for widget in (
-            self.repeat_hold_label,
-            self.repeat_hold_spin,
-            self.repeat_hold_ms_label,
-            self.repeat_wait_label,
-            self.repeat_wait_spin,
-            self.repeat_wait_ms_label,
-        ):
-            widget.set_visible(rf_active)
-
-    def _on_repeat_rapidfire_toggled(self, _check: Gtk.CheckButton) -> None:
-        self._update_repeat_rapidfire_visibility()
-
-    def _build_superkey_tab(self) -> Gtk.Widget:
-        outer = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=0)
-
-        toolbar_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        toolbar_row.set_margin_top(8)
-        toolbar_row.set_margin_bottom(4)
-        toolbar_row.set_margin_start(12)
-        toolbar_row.set_margin_end(12)
-        toolbar_row.set_halign(Gtk.Align.START)
-
-        manage_btn = Gtk.Button(label="Open Super Keys…")
-        manage_btn.add_css_class("flat")
-        manage_btn.set_tooltip_text("Create or edit super keys")
-        manage_btn.connect("clicked", self._on_open_superkey_manager_clicked)
-        toolbar_row.append(manage_btn)
-
-        selection_hint = Gtk.Label(label="Select one · right-click to edit")
-        selection_hint.add_css_class("dim-label")
-        selection_hint.add_css_class("caption")
-        selection_hint.set_halign(Gtk.Align.START)
-        toolbar_row.append(selection_hint)
-        outer.append(toolbar_row)
-
-        scrolled = Gtk.ScrolledWindow()
-        scrolled.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC)
-        scrolled.set_vexpand(True)
-
-        self._superkey_listbox = Gtk.ListBox()
-        self._superkey_listbox.set_selection_mode(Gtk.SelectionMode.SINGLE)
-        self._superkey_listbox.set_valign(Gtk.Align.START)
-        self._superkey_listbox.add_css_class("boxed-list")
-        self._superkey_listbox.set_margin_start(12)
-        self._superkey_listbox.set_margin_end(12)
-        self._superkey_listbox.connect("row-selected", self._on_superkey_row_selected)
-        scrolled.set_child(self._superkey_listbox)
-        outer.append(scrolled)
-
-        self._load_superkey_list()
-        return outer
-
-    def _on_open_superkey_manager_clicked(self, _button: Gtk.Button) -> None:
-        self._open_superkey_manager()
-
-    def _open_superkey_manager(self, select_name: str | None = None) -> None:
-        from keymasq.gui.widgets.superkey_dialog import SuperkeyDialog
-
-        root = self.get_root()
-        profile_manager = self._profile_manager_for_child_dialog()
-        dialog = SuperkeyDialog(root, profile_manager)
-        dialog.connect("superkey-saved", self._on_superkey_manager_changed)
-        dialog.connect("superkey-deleted", self._on_superkey_manager_changed)
-        dialog.present(root)
-        if select_name:
-            dialog.select_superkey_by_name(select_name)
-
-    def _on_superkey_manager_changed(self, _dialog, _name: str) -> None:
-        notify_session_reload_async()
-        self._load_superkey_list()
-
-    def _set_initial_analog_tab(self) -> None:
-        action = self._current_action
-        if action is not None and action.action_type == ActionType.ANALOG_CONTROL:
-            self.stack.set_visible_child_name("analog_control")
-            return
-        if action is not None and action.action_type in (
-            ActionType.SUPPRESS,
-            ActionType.PASSTHROUGH,
-        ):
-            self.stack.set_visible_child_name("special")
-            return
-        if self._analog_control_list:
-            self.stack.set_visible_child_name("analog_control")
-        else:
-            self.stack.set_visible_child_name("analog_presets")
-
-    def _build_analog_presets_tab(self) -> Gtk.Widget:
-        outer = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=0)
-
-        intro = Gtk.Label(
-            label="Pick what this control does — one click sets it up. "
-            "Fine-tune it any time under Analog Controls."
-        )
-        intro.add_css_class("dim-label")
-        intro.set_wrap(True)
-        intro.set_justify(Gtk.Justification.CENTER)
-        intro.set_halign(Gtk.Align.CENTER)
-        intro.set_margin_top(16)
-        intro.set_margin_start(16)
-        intro.set_margin_end(16)
-        intro.set_margin_bottom(12)
-        outer.append(intro)
-
-        scrolled = Gtk.ScrolledWindow()
-        scrolled.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC)
-        scrolled.set_vexpand(True)
-
-        flow = Gtk.FlowBox()
-        flow.set_selection_mode(Gtk.SelectionMode.NONE)
-        flow.set_max_children_per_line(2)
-        flow.set_min_children_per_line(2)
-        flow.set_homogeneous(True)
-        flow.set_column_spacing(12)
-        flow.set_row_spacing(12)
-        flow.set_margin_start(16)
-        flow.set_margin_end(16)
-        flow.set_margin_bottom(16)
-        flow.set_valign(Gtk.Align.START)
-
-        for preset in analog_control_presets(self._analog_input_type):
-            flow.append(self._build_analog_preset_card(preset))
-
-        scrolled.set_child(flow)
-        outer.append(scrolled)
-
-        link_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        link_row.set_halign(Gtk.Align.CENTER)
-        link_row.set_margin_top(4)
-        link_row.set_margin_bottom(12)
-        manage_btn = Gtk.Button(label="Open Analog Controls…")
-        manage_btn.add_css_class("flat")
-        manage_btn.set_tooltip_text("Create or fine-tune analog controls")
-        manage_btn.connect("clicked", self._on_open_analog_manager_clicked)
-        link_row.append(manage_btn)
-        outer.append(link_row)
-
-        return outer
-
-    def _build_analog_preset_card(self, preset: AnalogControlPreset) -> Gtk.Widget:
-        button = Gtk.Button()
-        button.add_css_class("card")
-        button.set_hexpand(True)
-        button.set_tooltip_text(preset.description)
-        button.connect("clicked", self._on_analog_preset_clicked, preset)
-
-        content = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=12)
-        content.set_margin_top(12)
-        content.set_margin_bottom(12)
-        content.set_margin_start(12)
-        content.set_margin_end(12)
-
-        icon = Gtk.Image.new_from_icon_name(preset.icon_name)
-        icon.set_pixel_size(28)
-        content.append(icon)
-
-        text_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=2)
-        text_box.set_valign(Gtk.Align.CENTER)
-        title = Gtk.Label(label=preset.label)
-        title.set_halign(Gtk.Align.START)
-        title.add_css_class("heading")
-        text_box.append(title)
-        subtitle = Gtk.Label(label=preset.description)
-        subtitle.set_halign(Gtk.Align.START)
-        subtitle.set_wrap(True)
-        subtitle.add_css_class("dim-label")
-        subtitle.add_css_class("caption")
-        text_box.append(subtitle)
-        content.append(text_box)
-
-        button.set_child(content)
-        return button
-
-    def _on_analog_preset_clicked(self, _button: Gtk.Button, preset: AnalogControlPreset) -> None:
-        manager = AnalogControlManager()
-        name = manager.unique_analog_control_name(preset.default_name)
-        config = preset.build(name)
-        try:
-            manager.save_analog_control(config)
-        except (OSError, ValueError):
-            log.exception("Could not save preset analog control %s", name)
-            return
-        notify_session_reload_async()
-        action = MappingAction(
-            action_type=ActionType.ANALOG_CONTROL,
-            analog_control_names=[name],
-        )
-        self.emit("key-selected", action)
-        self.close()
-
-    def _on_open_analog_manager_clicked(self, _button: Gtk.Button) -> None:
-        self._open_analog_control_manager()
-
-    def _open_analog_control_manager(self, select_name: str | None = None) -> None:
-        from keymasq.gui.widgets.analog_control_dialog import AnalogControlDialog
-
-        root = self.get_root()
-        profile_manager = self._profile_manager_for_child_dialog()
-        dialog = AnalogControlDialog(root, profile_manager)
-        dialog.connect("analog-control-saved", self._on_analog_control_manager_changed)
-        dialog.connect("analog-control-deleted", self._on_analog_control_manager_changed)
-        dialog.present(root)
-        if select_name:
-            dialog.select_control_by_name(select_name)
-
-    def _on_analog_control_manager_changed(self, _dialog, _name: str) -> None:
-        notify_session_reload_async()
-        self._load_analog_control_list()
-        if self.stack.get_visible_child_name() == "analog_presets":
-            self.stack.set_visible_child_name("analog_control")
-        self._on_tab_changed(self.stack, None)
-
-    def _build_analog_control_tab(self) -> Gtk.Widget:
-        outer = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=0)
-
-        toolbar_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        toolbar_row.set_margin_top(8)
-        toolbar_row.set_margin_bottom(4)
-        toolbar_row.set_margin_start(12)
-        toolbar_row.set_margin_end(12)
-        toolbar_row.set_halign(Gtk.Align.START)
-
-        manage_btn = Gtk.Button(label="Open Analog Controls…")
-        manage_btn.add_css_class("flat")
-        manage_btn.set_tooltip_text("Create or fine-tune analog controls")
-        manage_btn.connect("clicked", self._on_open_analog_manager_clicked)
-        toolbar_row.append(manage_btn)
-
-        selection_hint = Gtk.Label(
-            label="Select one or multiple · right-click to edit"
-        )
-        selection_hint.add_css_class("dim-label")
-        selection_hint.add_css_class("caption")
-        selection_hint.set_halign(Gtk.Align.START)
-        toolbar_row.append(selection_hint)
-        outer.append(toolbar_row)
-
-        scrolled = Gtk.ScrolledWindow()
-        scrolled.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC)
-        scrolled.set_vexpand(True)
-
-        self._analog_control_listbox = Gtk.ListBox()
-        self._analog_control_listbox.set_selection_mode(Gtk.SelectionMode.MULTIPLE)
-        self._analog_control_listbox.set_valign(Gtk.Align.START)
-        self._analog_control_listbox.add_css_class("boxed-list")
-        self._analog_control_listbox.set_margin_start(12)
-        self._analog_control_listbox.set_margin_end(12)
-        self._analog_control_listbox.connect(
-            "row-selected",
-            self._on_analog_control_row_selected,
-        )
-        scrolled.set_child(self._analog_control_listbox)
-        outer.append(scrolled)
-
-        self._load_analog_control_list()
-        return outer
-
-    def _build_keyboard_tab(self) -> Gtk.Widget:
-        outer = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=0)
-
-        scrolled = build_shared_keyboard_tab(
-            self,
-            keyboard_layout=KEYBOARD_LAYOUT,
-            key_to_evdev=KEY_TO_EVDEV,
-            key_widths=KEY_WIDTHS,
-            system_key_groups=SYSTEM_KEY_GROUPS,
-        )
-        scrolled.set_vexpand(True)
-        outer.append(scrolled)
-
-        toolbar_sep = Gtk.Separator(orientation=Gtk.Orientation.HORIZONTAL)
-        toolbar_sep.set_margin_start(12)
-        toolbar_sep.set_margin_end(12)
-        outer.append(toolbar_sep)
-
-        toolbar = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        toolbar.set_halign(Gtk.Align.CENTER)
-        toolbar.set_margin_top(12)
-        toolbar.set_margin_bottom(12)
-        toolbar.set_margin_start(12)
-        toolbar.set_margin_end(12)
-
-        self.kb_capture_btn = Gtk.Button(label="Capture Key")
-        self.kb_capture_btn.connect("clicked", self._on_keyboard_capture_clicked)
-        toolbar.append(self.kb_capture_btn)
-
-        self.kb_capture_status = Gtk.Label(label="")
-        self.kb_capture_status.add_css_class("dim-label")
-        toolbar.append(self.kb_capture_status)
-
-        toolbar.append(Gtk.Separator(orientation=Gtk.Orientation.VERTICAL))
-
-        toolbar.append(Gtk.Label(label="Key code:"))
-
-        self.kb_code_entry = Gtk.Entry()
-        self.kb_code_entry.set_placeholder_text("e.g. 125 or key_leftmeta")
-        self.kb_code_entry.set_width_chars(18)
-        toolbar.append(self.kb_code_entry)
-
-        code_btn = Gtk.Button(label="Map Code")
-        code_btn.connect("clicked", self._on_map_code_clicked)
-        toolbar.append(code_btn)
-
-        outer.append(toolbar)
-
-        if not hasattr(self, "_kb_capture_controller"):
-            self._kb_capture_pending = False
-            self._kb_capture_controller = Gtk.EventControllerKey()
-            self._kb_capture_controller.connect(
-                "key-pressed", self._on_keyboard_capture_key_pressed
-            )
-            self.add_controller(self._kb_capture_controller)
-
-        return outer
-
-    def _on_keyboard_capture_clicked(self, btn) -> None:
-        self._kb_capture_pending = True
-        self.kb_capture_status.set_text("Press a key...")
-
-    def _on_keyboard_capture_key_pressed(self, controller, keyval, keycode, state) -> bool:
-        if not getattr(self, "_kb_capture_pending", False):
-            return False
-        evdev_name = self._keyval_to_evdev(keyval)
-        if not evdev_name:
-            self.kb_capture_status.set_text("Unrecognized key")
-            self._kb_capture_pending = False
-            return True
-        self.kb_capture_status.set_text(f"Captured: {evdev_name}")
-        self._kb_capture_pending = False
-        self._emit_keyboard_mapping(evdev_name)
-        return True
-
-    def _on_map_code_clicked(self, btn) -> None:
-        raw = self.kb_code_entry.get_text().strip().lower()
-        if not raw:
-            return
-        evdev_name = None
-        if raw.startswith("key_"):
-            evdev_name = raw
-        else:
-            try:
-                code = int(raw)
-                key_name = evdev.ecodes.KEY.get(code)
-                if isinstance(key_name, str) and key_name.startswith("KEY_"):
-                    evdev_name = key_name.lower()
-            except (TypeError, ValueError):
-                evdev_name = None
-        if not evdev_name:
-            self.kb_code_entry.set_text("")
-            self.kb_code_entry.set_placeholder_text("Unknown key code")
-            return
-        self._emit_keyboard_mapping(evdev_name)
-
-    def _emit_keyboard_mapping(self, evdev_name: str) -> None:
-        use_rapidfire = _keyboard_target_allows_rapidfire(evdev_name)
-        use_tap = _keyboard_target_allows_tap(evdev_name)
-        action = MappingAction(
-            action_type=ActionType.KEYBOARD,
-            target=evdev_name,
-            rapidfire_enabled=self._rapidfire_enabled if use_rapidfire else False,
-            rapidfire_hold_ms=int(self.hold_spin.get_value()) if use_rapidfire else 20,
-            rapidfire_wait_ms=int(self.wait_spin.get_value()) if use_rapidfire else 20,
-            tap_enabled=self._tap_enabled if use_tap else False,
-            tap_hold_ms=int(self.tap_spin.get_value()) if use_tap else 150,
-        )
-        self.emit("key-selected", action)
-        self.close()
-
-    def _keyval_to_evdev(self, keyval: int) -> str | None:
-        name = (Gdk.keyval_name(keyval) or "").lower()
-        if not name:
-            return None
-        special = {
-            "escape": "key_esc",
-            "tab": "key_tab",
-            "return": "key_enter",
-            "backspace": "key_backspace",
-            "space": "key_space",
-            "shift_l": "key_leftshift",
-            "shift_r": "key_rightshift",
-            "control_l": "key_leftctrl",
-            "control_r": "key_rightctrl",
-            "alt_l": "key_leftalt",
-            "alt_r": "key_rightalt",
-            "super_l": "key_leftmeta",
-            "super_r": "key_rightmeta",
-            "menu": "key_menu",
-            "left": "key_left",
-            "right": "key_right",
-            "up": "key_up",
-            "down": "key_down",
-            "minus": "key_minus",
-            "equal": "key_equal",
-            "bracketleft": "key_leftbrace",
-            "bracketright": "key_rightbrace",
-            "backslash": "key_backslash",
-            "semicolon": "key_semicolon",
-            "apostrophe": "key_apostrophe",
-            "comma": "key_comma",
-            "period": "key_dot",
-            "slash": "key_slash",
-        }
-        if name in special:
-            return special[name]
-        if len(name) == 1 and name.isalpha():
-            return f"key_{name}"
-        if name.isdigit():
-            return f"key_{name}"
-        if name.startswith("f") and name[1:].isdigit():
-            return f"key_{name}"
-        return None
-
-    def _build_navigation_tab(self) -> Gtk.Widget:
-        return build_shared_navigation_tab(self, f_extra=F_EXTRA)
-
-    def _build_media_tab(self) -> Gtk.Widget:
-        return build_shared_media_tab(
-            self,
-            media_groups=MEDIA_KEY_GROUPS,
-            mpris_groups=MPRIS_MEDIA_GROUPS,
-        )
-
-    def _build_mouse_tab(self) -> Gtk.Widget:
-        box = build_shared_mouse_tab(self)
-        box.append(Gtk.Separator())
-
-        move_label = Gtk.Label(label="Move Cursor")
-        move_label.add_css_class("dim-label")
-        box.append(move_label)
-
-        mode_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        mode_row.set_halign(Gtk.Align.CENTER)
-
-        self.mouse_move_rel_check = Gtk.CheckButton(label="Relative")
-        self.mouse_move_rel_check.set_active(self._mouse_move_mode == "rel")
-        self.mouse_move_rel_check.connect("toggled", self._on_mouse_move_mode_changed)
-        mode_row.append(self.mouse_move_rel_check)
-
-        self.mouse_move_abs_check = Gtk.CheckButton(label="Absolute")
-        self.mouse_move_abs_check.set_group(self.mouse_move_rel_check)
-        self.mouse_move_abs_check.set_active(self._mouse_move_mode == "abs")
-        self.mouse_move_abs_check.connect("toggled", self._on_mouse_move_mode_changed)
-        mode_row.append(self.mouse_move_abs_check)
-
-        box.append(mode_row)
-
-        coords_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        coords_row.set_halign(Gtk.Align.CENTER)
-
-        x_label = Gtk.Label(label="X:")
-        coords_row.append(x_label)
-        self.mouse_move_x_spin = Gtk.SpinButton()
-        self.mouse_move_x_spin.set_adjustment(
-            Gtk.Adjustment(value=self._mouse_move_x, lower=-10000, upper=10000, step_increment=1)
-        )
-        self.mouse_move_x_spin.set_width_chars(6)
-        coords_row.append(self.mouse_move_x_spin)
-
-        y_label = Gtk.Label(label="Y:")
-        coords_row.append(y_label)
-        self.mouse_move_y_spin = Gtk.SpinButton()
-        self.mouse_move_y_spin.set_adjustment(
-            Gtk.Adjustment(value=self._mouse_move_y, lower=-10000, upper=10000, step_increment=1)
-        )
-        self.mouse_move_y_spin.set_width_chars(6)
-        coords_row.append(self.mouse_move_y_spin)
-
-        capture_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        capture_row.set_halign(Gtk.Align.CENTER)
-
-        if not self._slurp_available:
-            delay_label = Gtk.Label(label="Capture in:")
-            capture_row.append(delay_label)
-
-        self.mouse_move_capture_delay_spin = Gtk.SpinButton()
-        self.mouse_move_capture_delay_spin.set_adjustment(
-            Gtk.Adjustment(
-                value=self._capture_delay_seconds,
-                lower=0.2,
-                upper=15.0,
-                step_increment=0.2,
-            )
-        )
-        self.mouse_move_capture_delay_spin.set_digits(1)
-        self.mouse_move_capture_delay_spin.set_width_chars(4)
-        self.mouse_move_capture_delay_spin.set_visible(not self._slurp_available)
-        capture_row.append(self.mouse_move_capture_delay_spin)
-
-        if not self._slurp_available:
-            delay_suffix = Gtk.Label(label="s")
-            capture_row.append(delay_suffix)
-
-        btn_label = "Capture" if self._slurp_available else "Capture Position"
-        self.mouse_move_capture_btn = Gtk.Button(label=btn_label)
-        self.mouse_move_capture_btn.connect("clicked", self._on_capture_position_clicked)
-        capture_row.append(self.mouse_move_capture_btn)
-
-        self.mouse_move_capture_status = Gtk.Label(label="")
-        self.mouse_move_capture_status.add_css_class("dim-label")
-        self.mouse_move_capture_status.set_halign(Gtk.Align.START)
-        capture_row.append(self.mouse_move_capture_status)
-
-        self.mouse_move_capture_row = capture_row
-        box.append(self.mouse_move_capture_row)
-
-        move_map_btn = Gtk.Button(label="Map Move")
-        move_map_btn.add_css_class("suggested-action")
-        move_map_btn.connect("clicked", self._on_mouse_move_map_clicked)
-        coords_row.append(move_map_btn)
-
-        box.append(coords_row)
-
-        self._update_mouse_move_mode_visibility()
-
-        return box
-
-    def _build_gamepad_tab(self) -> Gtk.Widget:
-        box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=8)
-        warning = Gtk.Label(label="")
-        warning.set_margin_top(8)
-        warning.set_margin_start(12)
-        warning.set_margin_end(12)
-        warning.set_wrap(True)
-        warning.set_xalign(0)
-        warning.add_css_class("dim-label")
-        warning.add_css_class("warning")
-        self._gamepad_output_warning_label = warning
-        box.append(warning)
-        box.append(build_shared_gamepad_tab(self))
-        self._update_gamepad_output_warning()
-        self._prefill_gamepad_inputs()
-        return box
-
-    def _build_gamepad_output_header(self) -> Gtk.Widget | None:
-        choices = self._gamepad_output_choices()
-        if len(choices) <= 1 and not self._selected_gamepad_output_id:
-            return None
-
-        box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        box.set_valign(Gtk.Align.CENTER)
-        arrow = Gtk.Label(label="→")
-        arrow.add_css_class("dim-label")
-        self._gamepad_output_ids = [output_id for output_id, _label in choices]
-        dropdown = Gtk.DropDown.new_from_strings([label for _output_id, label in choices])
-        dropdown.set_valign(Gtk.Align.CENTER)
-        selected = 0
-        for index, output_id in enumerate(self._gamepad_output_ids):
-            if gamepad_output_choice_matches(output_id, self._selected_gamepad_output_id):
-                selected = index
-                break
-        dropdown.set_selected(selected)
-        dropdown.connect("notify::selected", self._on_gamepad_output_selected)
-        self._gamepad_output_dropdown = dropdown
-        box.append(arrow)
-        box.append(dropdown)
-        box.set_visible(False)
-        return box
-
-    def _gamepad_output_choices(self) -> list[tuple[str | None, str]]:
-        count = virtual_gamepad_count()
-        try:
-            hardware_configs = list(HardwareManager().list_hardware())
-        except (OSError, RuntimeError) as exc:
-            log.debug("Unable to load hardware configs for gamepad outputs: %s", exc)
-            hardware_configs = []
-        return gamepad_output_choices_for(
-            self._selected_gamepad_output_id,
-            count,
-            hardware_configs,
-        )
-
-    def _on_gamepad_output_selected(self, dropdown: Gtk.DropDown, _param) -> None:
-        selected = int(dropdown.get_selected())
-        if 0 <= selected < len(self._gamepad_output_ids):
-            self._selected_gamepad_output_id = self._gamepad_output_ids[selected]
-        self._update_gamepad_output_warning()
-
-    def _update_gamepad_output_warning(self) -> None:
-        label = self._gamepad_output_warning_label
-        if label is None:
-            return
-        message = gamepad_output_unavailable_message(
-            self._selected_gamepad_output_id,
-            virtual_gamepad_count(),
-        )
-        label.set_label(message or "")
-        label.set_visible(bool(message))
-
-    def _build_options_box(self) -> Gtk.Widget:
-        box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=8)
-
-        row1 = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        row1.set_halign(Gtk.Align.START)
-
-        self.rapidfire_check = Gtk.CheckButton(label="Rapidfire")
-        self.rapidfire_check.set_active(self._rapidfire_enabled)
-        self.rapidfire_check.set_tooltip_text(DEFAULT_RAPIDFIRE_TOOLTIP)
-        self.rapidfire_check.connect("toggled", self._on_rapidfire_toggled)
-        row1.append(self.rapidfire_check)
-
-        self.hold_label = Gtk.Label(label="Hold:")
-        row1.append(self.hold_label)
-
-        self.hold_spin = Gtk.SpinButton()
-        hold_adj = Gtk.Adjustment(
-            value=self._rapidfire_hold,
-            lower=MIN_RAPIDFIRE_HOLD_MS,
-            upper=1000,
-            step_increment=1,
-        )
-        self.hold_spin.set_adjustment(hold_adj)
-        row1.append(self.hold_spin)
-
-        self.hold_ms_label = Gtk.Label(label="ms")
-        row1.append(self.hold_ms_label)
-
-        self.wait_label = Gtk.Label(label="Wait:")
-        row1.append(self.wait_label)
-
-        self.wait_spin = Gtk.SpinButton()
-        wait_adj = Gtk.Adjustment(
-            value=self._rapidfire_wait,
-            lower=MIN_RAPIDFIRE_WAIT_MS,
-            upper=1000,
-            step_increment=1,
-        )
-        self.wait_spin.set_adjustment(wait_adj)
-        row1.append(self.wait_spin)
-
-        self.wait_ms_label = Gtk.Label(label="ms")
-        row1.append(self.wait_ms_label)
-
-        box.append(row1)
-
-        row2 = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        row2.set_halign(Gtk.Align.START)
-
-        self.tap_check = Gtk.CheckButton(label="Tap")
-        self.tap_check.set_active(self._tap_enabled)
-        self.tap_check.set_tooltip_text(
-            "Send the action as a quick tap when the button is released within the hold window"
-        )
-        self.tap_check.connect("toggled", self._on_tap_toggled)
-        row2.append(self.tap_check)
-
-        self.tap_hold_label = Gtk.Label(label="Hold:")
-        row2.append(self.tap_hold_label)
-
-        self.tap_spin = Gtk.SpinButton()
-        tap_adj = Gtk.Adjustment(value=self._tap_hold, lower=10, upper=500, step_increment=10)
-        self.tap_spin.set_adjustment(tap_adj)
-        row2.append(self.tap_spin)
-
-        self.tap_ms_label = Gtk.Label(label="ms")
-        row2.append(self.tap_ms_label)
-
-        box.append(row2)
-
-        self._update_options_visibility()
-
-        return box
-
-    def _update_options_visibility(self):
-        rf_active = self._allow_rapidfire and self.rapidfire_check.get_active()
-        tap_visible = self._allow_tap
-        tap_active = tap_visible and self.tap_check.get_active()
-
-        self.rapidfire_check.set_visible(self._allow_rapidfire)
-        self.tap_check.set_visible(tap_visible)
-
-        self.hold_label.set_visible(rf_active)
-        self.hold_spin.set_visible(rf_active)
-        self.hold_ms_label.set_visible(rf_active)
-        self.wait_label.set_visible(rf_active)
-        self.wait_spin.set_visible(rf_active)
-        self.wait_ms_label.set_visible(rf_active)
-
-        self.tap_hold_label.set_visible(tap_active)
-        self.tap_spin.set_visible(tap_active)
-        self.tap_ms_label.set_visible(tap_active)
-
-    def _create_key_button(
-        self, label: str, evdev: str, width: float = 1, large: bool = False, protected: bool = False
-    ) -> Gtk.Button:
-        btn = Gtk.Button(label=label)
-        btn.add_css_class("key-button")
-
-        if large:
-            btn.set_size_request(200, 50)
-        else:
-            base_width = 36
-            btn.set_size_request(int(base_width * width), 34)
-
-        if protected:
-            btn.add_css_class("protected-key")
-            btn.set_tooltip_text("Protected - cannot remap")
-
-        btn._evdev_name = evdev
-        btn._protected = protected
-        return btn
-
-    def _on_tab_changed(self, stack, param):
-        child_name = self.stack.get_visible_child_name()
-        is_special = child_name == "special"
-        is_superkey = child_name == "superkey"
-        is_analog_control = child_name == "analog_control"
-        is_analog_presets = child_name == "analog_presets"
-        is_analog_only = is_analog_control or is_analog_presets
-        is_macro = child_name == "macro"
-        is_profile = child_name == "profile"
-        is_exec = child_name == "exec"
-        is_gamepad = child_name == "gamepad"
-        is_media = child_name == "media"
-        is_compositor_action = child_name in self._compositor_action_page_ids
-        has_options = self._allow_rapidfire or self._allow_tap
-        # Repeat carries its own rapidfire controls in the Special tab, so the
-        # shared footer options only apply to the concrete-input tabs.
-        show_options = has_options and not (
-            is_special
-            or is_superkey
-            or is_analog_only
-            or is_macro
-            or is_profile
-            or is_exec
-            or is_compositor_action
-            or is_media
-        )
-        self.options_box.set_sensitive(show_options)
-        self.options_box.set_visible(show_options)
-        self._update_options_visibility()
-        self.map_btn.set_visible(is_superkey or is_analog_control or is_macro or is_profile)
-        if self._cancel_macro_playback_btn is not None:
-            self._cancel_macro_playback_btn.set_visible(is_macro)
-        if is_superkey:
-            self.map_btn.set_sensitive(self._selected_superkey is not None)
-        elif is_analog_control:
-            self.map_btn.set_sensitive(bool(self._selected_analog_controls))
-        elif is_macro:
-            self.map_btn.set_sensitive(self._selected_macro is not None)
-        elif is_profile:
-            self.map_btn.set_sensitive(bool(self._selected_profile_name))
-        else:
-            self.map_btn.set_sensitive(False)
-        if self._gamepad_output_header is not None:
-            self._gamepad_output_header.set_visible(is_gamepad)
-        self._update_actions_docs_button()
-
-    def _active_actions_docs_link(self) -> tuple[str, str] | None:
-        child_name = self.stack.get_visible_child_name()
-        if not child_name:
-            return None
-        return ACTION_DOC_LINKS.get(child_name)
-
-    def _update_actions_docs_button(self) -> None:
-        if not hasattr(self, "actions_docs_btn"):
-            return
-        link = self._active_actions_docs_link()
-        self.actions_docs_btn.set_visible(link is not None)
-        if link is None:
-            return
-        _anchor, title = link
-        self.actions_docs_btn.set_tooltip_text(f"Open {title} documentation")
-
-    def _on_actions_docs_clicked(self, _button: Gtk.Button) -> None:
-        link = self._active_actions_docs_link()
-        if link is None:
-            return
-        anchor, _title = link
-        url = _actions_docs_url(anchor)
-        try:
-            launcher = Gtk.UriLauncher.new(url)
-            launcher.launch(None, None, None)
-        except Exception:
-            log.exception("Could not open action documentation %s", url)
-
-    def _warn_and_clear_unsupported_rapidfire(self, action_type: ActionType) -> None:
-        if not self._rapidfire_enabled or action_type_supports_rapidfire(action_type):
-            return
-        log.warning(
-            "Ignoring rapidfire for unsupported %s action in key selector",
-            action_type.value,
-        )
-        if self.rapidfire_check.get_active():
-            self.rapidfire_check.set_active(False)
-        else:
-            self._rapidfire_enabled = False
-            self._update_options_visibility()
-
-    def _on_rapidfire_toggled(self, check):
-        if not self._allow_rapidfire:
-            return
-        self._rapidfire_enabled = check.get_active()
-        if self._rapidfire_enabled:
-            self.tap_check.set_active(False)
-            self._tap_enabled = False
-        self._update_options_visibility()
-
-    def _on_tap_toggled(self, check):
-        if not self._allow_tap:
-            return
-        self._tap_enabled = check.get_active()
-        if self._tap_enabled:
-            self.rapidfire_check.set_active(False)
-            self._rapidfire_enabled = False
-        self._update_options_visibility()
-
-    def _on_special_clicked(self, btn, action_type: str):
-        if action_type == "clear_mapping":
-            self.emit("key-selected", None)
-        elif action_type == "explicit_passthrough":
-            self._warn_and_clear_unsupported_rapidfire(ActionType.PASSTHROUGH)
-            action = MappingAction(action_type=ActionType.PASSTHROUGH)
-            self.emit("key-selected", action)
-        elif action_type == "suppress":
-            self._warn_and_clear_unsupported_rapidfire(ActionType.SUPPRESS)
-            action = MappingAction(action_type=ActionType.SUPPRESS)
-            self.emit("key-selected", action)
-        elif action_type == "cancel_macro_playback":
-            self._warn_and_clear_unsupported_rapidfire(ActionType.CANCEL_MACRO_PLAYBACK)
-            action = MappingAction(action_type=ActionType.CANCEL_MACRO_PLAYBACK)
-            self.emit("key-selected", action)
-        self.close()
-
-    def _on_repeat_button_toggled(self, btn: Gtk.ToggleButton) -> None:
-        if self._repeat_options_box is not None:
-            self._repeat_options_box.set_visible(btn.get_active())
-
-    def _on_repeat_category_toggled(
-        self,
-        check: Gtk.ToggleButton,
-        category: str,
-    ) -> None:
-        if check.get_active():
-            if category not in self._repeat_categories:
-                self._repeat_categories.append(category)
-        else:
-            self._repeat_categories = [
-                existing for existing in self._repeat_categories if existing != category
-            ]
-        self._update_repeat_map_button()
-
-    def _update_repeat_map_button(self) -> None:
-        if self._repeat_map_btn is not None:
-            self._repeat_map_btn.set_sensitive(bool(self._repeat_categories))
-
-    def _on_repeat_map_clicked(self, _btn: Gtk.Button) -> None:
-        if not self._repeat_categories:
-            return
-        rapidfire_enabled = bool(
-            self._allow_rapidfire and self.repeat_rapidfire_check.get_active()
-        )
-        action = MappingAction(
-            action_type=ActionType.REPEAT,
-            repeat_categories=list(self._repeat_categories),
-            rapidfire_enabled=rapidfire_enabled,
-            rapidfire_hold_ms=int(self.repeat_hold_spin.get_value()),
-            rapidfire_wait_ms=int(self.repeat_wait_spin.get_value()),
-        )
-        self.emit("key-selected", action)
-        self.close()
-
-    def _on_exec_text_changed(self, entry: Gtk.Entry) -> None:
-        self.exec_map_btn.set_sensitive(bool(entry.get_text().strip()))
-
-    def _on_exec_map_clicked(self, btn: Gtk.Button) -> None:
-        cmd = self.exec_entry.get_text().strip()
-        if not cmd:
-            return
-        self._warn_and_clear_unsupported_rapidfire(ActionType.EXEC)
-        action = MappingAction(action_type=ActionType.EXEC, cmd=cmd)
-        self.emit("key-selected", action)
-        self.close()
-
-    def _on_compositor_action_selected(self, action: MappingAction) -> None:
-        self._warn_and_clear_unsupported_rapidfire(ActionType.COMPOSITOR_DISPATCH)
-        self.emit("key-selected", action)
-        self.close()
-
-    def _on_mpris_clicked(self, btn, command: str) -> None:
-        self._warn_and_clear_unsupported_rapidfire(ActionType.MPRIS)
-        action = MappingAction(action_type=ActionType.MPRIS, mpris_command=command)
-        self.emit("key-selected", action)
-        self.close()
-
-    def _on_keyboard_clicked(self, btn, evdev_name: str):
-        use_rapidfire = _keyboard_target_allows_rapidfire(evdev_name)
-        use_tap = _keyboard_target_allows_tap(evdev_name)
-        action = MappingAction(
-            action_type=ActionType.KEYBOARD,
-            target=evdev_name,
-            rapidfire_enabled=self._rapidfire_enabled if use_rapidfire else False,
-            rapidfire_hold_ms=int(self.hold_spin.get_value()) if use_rapidfire else 20,
-            rapidfire_wait_ms=int(self.wait_spin.get_value()) if use_rapidfire else 20,
-            tap_enabled=self._tap_enabled if use_tap else False,
-            tap_hold_ms=int(self.tap_spin.get_value()) if use_tap else 150,
-        )
-        self.emit("key-selected", action)
-        self.close()
-
-    def _on_f_key_selected(self, btn):
-        idx = self.f_dropdown.get_selected()
-        f_key = F_EXTRA[idx]
-        evdev_name = KEY_TO_EVDEV.get(f_key)
-        if evdev_name:
-            self._on_keyboard_clicked(btn, evdev_name)
-
-    def _on_mouse_clicked(self, btn, evdev_name: str):
-        action = MappingAction(
-            action_type=ActionType.MOUSE,
-            target=evdev_name,
-            rapidfire_enabled=self._rapidfire_enabled,
-            rapidfire_hold_ms=int(self.hold_spin.get_value()),
-            rapidfire_wait_ms=int(self.wait_spin.get_value()),
-            tap_enabled=self._tap_enabled,
-            tap_hold_ms=int(self.tap_spin.get_value()),
-        )
-        self.emit("key-selected", action)
-        self.close()
-
-    def _on_mouse_move_map_clicked(self, btn) -> None:
-        x = int(self.mouse_move_x_spin.get_value())
-        y = int(self.mouse_move_y_spin.get_value())
-        if self.mouse_move_abs_check.get_active():
-            action_type = ActionType.MOUSE_MOVE_ABS
-        else:
-            action_type = ActionType.MOUSE_MOVE_REL
-        action = MappingAction(
-            action_type=action_type,
-            move_x=x,
-            move_y=y,
-            rapidfire_enabled=self._rapidfire_enabled,
-            rapidfire_hold_ms=int(self.hold_spin.get_value()),
-            rapidfire_wait_ms=int(self.wait_spin.get_value()),
-            tap_enabled=self._tap_enabled,
-            tap_hold_ms=int(self.tap_spin.get_value()),
-        )
-        self.emit("key-selected", action)
-        self.close()
-
-    def _on_mouse_move_mode_changed(self, check: Gtk.CheckButton) -> None:
-        self._update_mouse_move_mode_visibility()
-
-    def _update_mouse_move_mode_visibility(self) -> None:
-        is_abs = self.mouse_move_abs_check.get_active()
-        self.mouse_move_capture_row.set_visible(is_abs)
-        if not is_abs:
-            self._cancel_capture_position("")
-
-    def _on_capture_position_clicked(self, btn: Gtk.Button) -> None:
-        self._begin_position_capture(
-            self.mouse_move_capture_btn,
-            self.mouse_move_capture_status,
-            self._apply_mouse_move_capture_position,
-        )
-
-    def _apply_mouse_move_capture_position(self, x: int, y: int) -> None:
-        self.mouse_move_x_spin.set_value(int(x))
-        self.mouse_move_y_spin.set_value(int(y))
-
-    def _capture_compositor_position(
-        self,
-        button: Gtk.Button,
-        status_label: Gtk.Label,
-        callback: Callable[[int, int], None],
-    ) -> None:
-        self._begin_position_capture(
-            button,
-            status_label,
-            callback,
-        )
-
-    def _begin_position_capture(
-        self,
-        button: Gtk.Button | None,
-        status_label: Gtk.Label | None,
-        apply_position: Callable[[int, int], None],
-    ) -> None:
-        self._cancel_capture_position("")
-        self._capture_request_id += 1
-        request_id = self._capture_request_id
-        self._capture_button = button
-        self._capture_status_label = status_label
-        self._capture_apply = apply_position
-
-        if self._slurp_available:
-            self._capture_pending = True
-            if button is not None:
-                button.set_sensitive(False)
-            if status_label is not None:
-                status_label.set_text("Click to capture position...")
-            self._slurp_capture.capture_point(
-                lambda result, expected_id=request_id: self._on_slurp_capture_result(
-                    expected_id, result
-                )
-            )
-        else:
-            self._capture_delay_seconds = float(self.mouse_move_capture_delay_spin.get_value())
-            self._capture_pending = True
-            if button is not None:
-                button.set_sensitive(False)
-            if status_label is not None:
-                status_label.set_text(
-                    f"Move cursor now... capturing in {self._capture_delay_seconds:.1f}s"
-                )
-            self._capture_timeout_id = GLib.timeout_add(
-                int(self._capture_delay_seconds * 1000),
-                lambda expected_id=request_id: self._capture_position_after_delay(expected_id),
-            )
-
-    def _on_slurp_capture_result(self, request_id: int, result) -> None:
-        if request_id != self._capture_request_id:
-            return
-        self._capture_pending = False
-        if self._capture_button is not None:
-            self._capture_button.set_sensitive(True)
-
-        if result is None:
-            if self._capture_status_label is not None:
-                self._capture_status_label.set_text("Capture cancelled or failed")
-            return
-
-        if self._capture_apply is not None:
-            self._capture_apply(int(result.x), int(result.y))
-        if self._capture_status_label is not None:
-            self._capture_status_label.set_text(f"Captured: {result.x}, {result.y}")
-
-    def _capture_position_after_delay(self, request_id: int) -> bool:
-        self._capture_timeout_id = 0
-        if request_id != self._capture_request_id or not self._capture_pending:
-            return False
-        if self._capture_status_label is not None:
-            self._capture_status_label.set_text("Reading cursor position...")
-        session_request_async(
-            {"command": "get_cursor_position"},
-            lambda response, expected_id=request_id: self._on_capture_position_response(
-                expected_id, response
-            ),
-            timeout=5.0,
-        )
-        return False
-
-    def _on_capture_position_response(self, request_id: int, response: dict | None) -> bool:
-        if request_id != self._capture_request_id:
-            return False
-        self._capture_pending = False
-        if self._capture_button is not None:
-            self._capture_button.set_sensitive(True)
-
-        if not response or response.get("status") != "ok":
-            message = (
-                (response or {}).get("message") or (response or {}).get("error") or "Capture failed"
-            )
-            if "Unknown command: get_cursor_position" in message:
-                message = "Please restart Keymasq Session, then try again"
-            if self._capture_status_label is not None:
-                self._capture_status_label.set_text(message)
-            return False
-
-        x = int(response.get("x", 0))
-        y = int(response.get("y", 0))
-        if self._capture_apply is not None:
-            self._capture_apply(x, y)
-        if self._capture_status_label is not None:
-            self._capture_status_label.set_text("Captured")
-        return False
-
-    def _cancel_capture_position(self, status_text: str) -> None:
-        self._capture_request_id += 1
-        if self._capture_timeout_id:
-            GLib.source_remove(self._capture_timeout_id)
-            self._capture_timeout_id = 0
-        self._capture_pending = False
-        if self._capture_button is not None:
-            self._capture_button.set_sensitive(True)
-        if self._capture_status_label is not None:
-            self._capture_status_label.set_text(status_text)
-        self._capture_apply = None
-        self._capture_button = None
-        self._capture_status_label = None
-
-    def close(self) -> None:
-        self._cancel_capture_position("")
-        super().close()
-
-    def _on_gamepad_clicked(self, btn, evdev_name: str):
-        action = MappingAction(
-            action_type=ActionType.GAMEPAD,
-            target=evdev_name,
-            output_id=self._selected_gamepad_output_id,
-            rapidfire_enabled=self._rapidfire_enabled,
-            rapidfire_hold_ms=int(self.hold_spin.get_value()),
-            rapidfire_wait_ms=int(self.wait_spin.get_value()),
-            tap_enabled=self._tap_enabled,
-            tap_hold_ms=int(self.tap_spin.get_value()),
-        )
-        self.emit("key-selected", action)
-        self.close()
-
-    def _on_gamepad_axis_clicked(self, btn, axis_target: str, axis_value: int):
-        action = MappingAction(
-            action_type=ActionType.GAMEPAD_AXIS,
-            target=axis_target,
-            axis_value=int(axis_value),
-            output_id=self._selected_gamepad_output_id,
-            rapidfire_enabled=self._rapidfire_enabled,
-            rapidfire_hold_ms=int(self.hold_spin.get_value()),
-            rapidfire_wait_ms=int(self.wait_spin.get_value()),
-            tap_enabled=self._tap_enabled,
-            tap_hold_ms=int(self.tap_spin.get_value()),
-        )
-        self.emit("key-selected", action)
-        self.close()
-
-    def _on_gamepad_code_clicked(self, widget) -> None:
-        evdev_name = _resolve_gamepad_button_target(self.gamepad_code_entry.get_text())
-        if not evdev_name:
-            self.gamepad_code_entry.set_text("")
-            self.gamepad_code_entry.set_placeholder_text("Unknown button code")
-            return
-        self._on_gamepad_clicked(widget, evdev_name)
-
-    def _build_macro_tab(self) -> Gtk.Widget:
-        outer = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=0)
-
-        outer.append(self._build_macro_slot_console())
-
-        outer.append(Gtk.Separator())
-
-        library_label = Gtk.Label(label="Macro Library")
-        library_label.add_css_class("dim-label")
-        library_label.set_halign(Gtk.Align.START)
-        library_label.set_margin_top(8)
-        library_label.set_margin_bottom(6)
-        library_label.set_margin_start(12)
-        outer.append(library_label)
-
-        self._macro_search_entry = Gtk.SearchEntry()
-        self._macro_search_entry.set_placeholder_text("Search macros")
-        self._macro_search_entry.set_tooltip_text(
-            "Filter macros by name, device type, or event count"
-        )
-        self._macro_search_entry.set_margin_start(12)
-        self._macro_search_entry.set_margin_end(12)
-        self._macro_search_entry.set_margin_bottom(8)
-        outer.append(self._macro_search_entry)
-
-        scrolled = Gtk.ScrolledWindow()
-        scrolled.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC)
-        scrolled.set_vexpand(True)
-
-        self._macro_listbox = Gtk.ListBox()
-        self._macro_listbox.set_selection_mode(Gtk.SelectionMode.SINGLE)
-        self._macro_listbox.set_valign(Gtk.Align.START)
-        self._macro_listbox.add_css_class("boxed-list")
-        self._macro_listbox.set_margin_start(12)
-        self._macro_listbox.set_margin_end(12)
-        self._macro_listbox.connect("row-selected", self._on_macro_row_selected)
-        install_listbox_fuzzy_filter(
-            self._macro_listbox,
-            self._macro_search_entry,
-            after_filter_changed=self._after_macro_search_filter_changed,
-        )
-        scrolled.set_child(self._macro_listbox)
-        outer.append(scrolled)
-
-        self._macro_options_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=8)
-        self._macro_options_box.set_margin_top(8)
-        self._macro_options_box.set_margin_start(12)
-        self._macro_options_box.set_margin_end(12)
-        self._macro_options_box.set_visible(False)
-
-        opt_row1 = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=16)
-        opt_row1.set_halign(Gtk.Align.START)
-
-        self._macro_movement_check = Gtk.CheckButton(label="Replay mouse movement")
-        self._macro_movement_check.set_active(self._macro_replay_movement)
-        self._macro_movement_check.connect("toggled", self._on_macro_movement_toggled)
-        opt_row1.append(self._macro_movement_check)
-
-        self._macro_clicks_check = Gtk.CheckButton(label="Replay mouse clicks")
-        self._macro_clicks_check.set_active(self._macro_replay_clicks)
-        self._macro_clicks_check.connect("toggled", self._on_macro_clicks_toggled)
-        opt_row1.append(self._macro_clicks_check)
-
-        self._macro_options_box.append(opt_row1)
-
-        opt_row2 = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        opt_row2.set_halign(Gtk.Align.START)
-
-        speed_label = Gtk.Label(label="Speed:")
-        opt_row2.append(speed_label)
-
-        self._macro_speed_spin = Gtk.SpinButton()
-        speed_adj = Gtk.Adjustment(
-            value=self._macro_speed, lower=0.1, upper=10.0, step_increment=0.1
-        )
-        self._macro_speed_spin.set_adjustment(speed_adj)
-        self._macro_speed_spin.set_digits(1)
-        self._macro_speed_spin.connect("value-changed", self._on_macro_speed_changed)
-        opt_row2.append(self._macro_speed_spin)
-
-        speed_suffix = Gtk.Label(label="×")
-        opt_row2.append(speed_suffix)
-
-        self._macro_options_box.append(opt_row2)
-        outer.append(self._macro_options_box)
-
-        GLib.idle_add(self._load_macro_list)
-        return outer
-
-    def _build_macro_slot_console(self) -> Gtk.Widget:
-        """Slot-based recording/playback console.
-
-        Renders recording and playback controls for every temporary slot.
-        """
-        console = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=6)
-        console.set_margin_top(10)
-        console.set_margin_bottom(10)
-        console.set_margin_start(12)
-        console.set_margin_end(12)
-        self._macro_slot_console = console
-        self._refresh_macro_slot_console()
-
-        return console
-
-    def _refresh_macro_slot_console(self) -> None:
-        console = self._macro_slot_console
-        if console is None:
-            return
-        child = console.get_first_child()
-        while child is not None:
-            console.remove(child)
-            child = console.get_first_child()
-        self._macro_recording_enabled = self._resolve_macro_recording_enabled(
-            default=False
-        )
-        if self._macro_recording_enabled:
-            console.append(self._build_macro_slot_cards())
-        else:
-            console.append(self._build_macro_recording_disabled_placeholder())
-
-    def _build_macro_recording_disabled_placeholder(self) -> Gtk.Widget:
-        placeholder = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=10)
-        placeholder.add_css_class("card")
-        placeholder.add_css_class("macro-recording-disabled-placeholder")
-        placeholder.set_valign(Gtk.Align.CENTER)
-        placeholder.set_halign(Gtk.Align.FILL)
-
-        icon = Gtk.Image.new_from_icon_name("channel-insecure-symbolic")
-        icon.set_valign(Gtk.Align.START)
-        placeholder.append(icon)
-
-        text_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=2)
-        text_box.set_hexpand(True)
-        title = Gtk.Label(label="Macro recording is disabled")
-        title.add_css_class("caption-heading")
-        title.set_halign(Gtk.Align.START)
-        text_box.append(title)
-
-        body = Gtk.Label(label="Enable it in Settings > Macro recording to bind slot actions.")
-        body.add_css_class("dim-label")
-        body.set_wrap(True)
-        body.set_halign(Gtk.Align.START)
-        text_box.append(body)
-        placeholder.append(text_box)
-
-        settings_content = Adw.ButtonContent(
-            icon_name="emblem-system-symbolic",
-            label="Open Settings",
-        )
-        settings_btn = Gtk.Button()
-        settings_btn.set_child(settings_content)
-        settings_btn.set_valign(Gtk.Align.CENTER)
-        settings_btn.set_tooltip_text("Open macro recording settings")
-        settings_btn.connect("clicked", self._present_macro_recording_settings)
-        settings_btn.set_sensitive(
-            any(
-                callable(getattr(candidate, "present_recording_settings_dialog", None))
-                for candidate in self._macro_parent_candidates()
-            )
-        )
-        placeholder.append(settings_btn)
-        return placeholder
-
-    def _build_macro_slot_cards(self) -> Gtk.Widget:
-        grid = Gtk.Grid(column_spacing=6, row_spacing=6)
-        grid.set_halign(Gtk.Align.CENTER)
-
-        for index in range(MAX_MACRO_RECORDING_SLOTS):
-            slot = index + 1
-            grid.attach(self._build_macro_slot_card(slot), index, 0, 1, 1)
-
-        return grid
-
-    def _build_macro_slot_card(self, slot: int) -> Gtk.Widget:
-        card = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=4)
-        card.add_css_class("card")
-        card.add_css_class("macro-slot-card")
-
-        card.append(self._build_macro_slot_header(f"Slot {slot}"))
-
-        buttons = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=4)
-        buttons.set_halign(Gtk.Align.CENTER)
-        buttons.append(self._create_record_slot_button(slot))
-        buttons.append(self._create_play_slot_button(slot))
-        card.append(buttons)
-        return card
-
-    def _build_macro_slot_header(self, label: str) -> Gtk.Widget:
-        header = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=4)
-        header.add_css_class("macro-slot-header")
-        header.set_halign(Gtk.Align.CENTER)
-
-        title = Gtk.Label(label=label)
-        title.add_css_class("caption-heading")
-        title.add_css_class("macro-slot-title")
-        title.set_halign(Gtk.Align.CENTER)
-        header.append(title)
-        return header
-
-    def _create_record_slot_button(self, slot: int) -> Gtk.Button:
-        content = Adw.ButtonContent(icon_name="media-record-symbolic", label=str(slot))
-        content.add_css_class("macro-record-icon")
-        button = Gtk.Button()
-        button.add_css_class("macro-slot-button")
-        button.set_child(content)
-        button.set_size_request(42, 34)
-        button.set_tooltip_text(f"Toggle macro recording into slot {slot}")
-        button.connect("clicked", self._on_macro_recording_slot_clicked, slot)
-        if (
-            self._current_action
-            and self._current_action.action_type
-            in (ActionType.START_MACRO_RECORDING, ActionType.STOP_MACRO_RECORDING)
-            and self._current_action.macro_recording_slot == slot
-        ):
-            button.add_css_class("suggested-action")
-        return button
-
-    def _create_play_slot_button(self, slot: int) -> Gtk.Button:
-        content = Adw.ButtonContent(
-            icon_name="media-playback-start-symbolic", label=str(slot)
-        )
-        content.add_css_class("macro-play-icon")
-        button = Gtk.Button()
-        button.add_css_class("macro-slot-button")
-        button.set_child(content)
-        button.set_size_request(42, 34)
-        button.set_tooltip_text(f"Play the macro recorded in slot {slot}")
-        button.connect("clicked", self._on_macro_play_slot_clicked, slot)
-        if (
-            self._current_action
-            and self._current_action.action_type == ActionType.PLAY_MACRO_SLOT
-            and self._current_action.macro_recording_slot == slot
-        ):
-            button.add_css_class("suggested-action")
-        return button
-
-    def _build_profile_tab(self) -> Gtk.Widget:
-        outer = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=12)
-        outer.set_margin_top(12)
-        outer.set_margin_bottom(12)
-        outer.set_margin_start(12)
-        outer.set_margin_end(12)
-
-        title = Gtk.Label(label="Profile Controls")
-        title.add_css_class("title-4")
-        title.set_halign(Gtk.Align.START)
-        outer.append(title)
-
-        subtitle = Gtk.Label(label="Trigger global profile enable/disable/toggle.")
-        subtitle.add_css_class("dim-label")
-        subtitle.set_wrap(True)
-        subtitle.set_halign(Gtk.Align.START)
-        outer.append(subtitle)
-
-        action_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        action_row.set_halign(Gtk.Align.START)
-
-        action_label = Gtk.Label(label="Action")
-        action_label.set_size_request(90, -1)
-        action_label.set_halign(Gtk.Align.START)
-        action_row.append(action_label)
-
-        self._profile_action_dropdown = Gtk.DropDown()
-        action_model = Gtk.StringList()
-        action_model.append("Enable")
-        action_model.append("Toggle")
-        action_model.append("Disable")
-        self._profile_action_dropdown.set_model(action_model)
-        self._profile_action_dropdown.set_selected(
-            {"enable": 0, "toggle": 1, "disable": 2}.get(self._selected_profile_action, 0)
-        )
-        self._profile_action_dropdown.connect("notify::selected", self._on_profile_action_changed)
-        action_row.append(self._profile_action_dropdown)
-        outer.append(action_row)
-
-        profile_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        profile_row.set_halign(Gtk.Align.START)
-
-        profile_label = Gtk.Label(label="Profile")
-        profile_label.set_size_request(90, -1)
-        profile_label.set_halign(Gtk.Align.START)
-        profile_row.append(profile_label)
-
-        self._profile_name_dropdown = Gtk.DropDown()
-        self._profile_name_model = Gtk.StringList()
-        self._profile_name_dropdown.set_model(self._profile_name_model)
-        self._profile_name_dropdown.set_size_request(360, -1)
-        self._profile_name_dropdown.connect("notify::selected", self._on_profile_name_changed)
-        profile_row.append(self._profile_name_dropdown)
-        outer.append(profile_row)
-
-        self._profile_lifetime_box = self._build_profile_lifetime_controls()
-        outer.append(self._profile_lifetime_box)
-
-        self._profile_hint_label = Gtk.Label(label="")
-        self._profile_hint_label.add_css_class("dim-label")
-        self._profile_hint_label.set_halign(Gtk.Align.START)
-        self._profile_hint_label.set_wrap(True)
-        outer.append(self._profile_hint_label)
-
-        GLib.idle_add(self._load_profile_overview)
-        return outer
-
-    def _load_profile_overview(self) -> bool:
-        session_request_async({"command": "list_profiles"}, self._on_profile_overview_loaded)
-        return False
-
-    def _on_profile_overview_loaded(self, result: dict | None) -> bool:
-        result = result or {}
-        self._profile_entries = result.get("profiles", []) if result.get("status") == "ok" else []
-        self._populate_profile_names()
-        return False
-
-    def _on_profile_action_changed(self, dropdown, _pspec) -> None:
-        idx = dropdown.get_selected()
-        if idx == 0:
-            self._selected_profile_action = "enable"
-        elif idx == 2:
-            self._selected_profile_action = "disable"
-        else:
-            self._selected_profile_action = "toggle"
-        self._profile_lifetime_preset = self._coerce_profile_lifetime_preset(
-            self._profile_lifetime_preset
-        )
-        self._update_profile_lifetime_visibility()
-        self._update_profile_hint()
-
-    def _build_profile_lifetime_controls(self) -> Gtk.Widget:
-        box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=8)
-        box.set_halign(Gtk.Align.START)
-        self._profile_lifetime_box = box
-
-        self._profile_lifetime_title = Gtk.Label(label="Activation")
-        self._profile_lifetime_title.set_halign(Gtk.Align.START)
-        box.append(self._profile_lifetime_title)
-
-        preset_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        preset_label = Gtk.Label(label="Mode")
-        preset_label.set_size_request(90, -1)
-        preset_label.set_halign(Gtk.Align.START)
-        preset_row.append(preset_label)
-
-        self._profile_lifetime_dropdown = Gtk.DropDown()
-        self._profile_lifetime_model = Gtk.StringList()
-        self._sync_profile_lifetime_model()
-        self._profile_lifetime_dropdown.set_model(self._profile_lifetime_model)
-        self._profile_lifetime_dropdown.set_selected(
-            self._profile_lifetime_preset_index(self._profile_lifetime_preset)
-        )
-        self._profile_lifetime_dropdown.connect(
-            "notify::selected",
-            self._on_profile_lifetime_preset_changed,
-        )
-        preset_row.append(self._profile_lifetime_dropdown)
-        box.append(preset_row)
-
-        self._profile_lifetime_timeout_row = Gtk.Box(
-            orientation=Gtk.Orientation.HORIZONTAL,
-            spacing=8,
-        )
-        timeout_label = Gtk.Label(label="Timeout")
-        timeout_label.set_size_request(90, -1)
-        timeout_label.set_halign(Gtk.Align.START)
-        self._profile_lifetime_timeout_row.append(timeout_label)
-        self._profile_lifetime_timeout_adjustment = Gtk.Adjustment(
-            value=self._profile_lifetime_timeout_ms,
-            lower=1,
-            upper=3_600_000,
-            step_increment=50,
-        )
-        self._profile_lifetime_timeout_spin = Gtk.SpinButton()
-        self._profile_lifetime_timeout_spin.set_adjustment(
-            self._profile_lifetime_timeout_adjustment
-        )
-        self._profile_lifetime_timeout_spin.connect(
-            "value-changed",
-            self._on_profile_lifetime_timeout_changed,
-        )
-        self._profile_lifetime_timeout_row.append(self._profile_lifetime_timeout_spin)
-        timeout_unit = Gtk.Label(label="ms")
-        timeout_unit.add_css_class("dim-label")
-        self._profile_lifetime_timeout_row.append(timeout_unit)
-        box.append(self._profile_lifetime_timeout_row)
-
-        self._profile_lifetime_custom_box = Gtk.Box(
-            orientation=Gtk.Orientation.VERTICAL,
-            spacing=6,
-        )
-
-        self._profile_custom_count_row = Gtk.Box(
-            orientation=Gtk.Orientation.HORIZONTAL,
-            spacing=8,
-        )
-        self._profile_custom_count_row.set_halign(Gtk.Align.START)
-        self._profile_custom_count_toggle = Gtk.ToggleButton(label="Action count")
-        self._profile_custom_count_toggle.set_size_request(140, -1)
-        self._profile_custom_count_toggle.set_active(self._profile_custom_action_count)
-        self._profile_custom_count_toggle.connect(
-            "toggled",
-            self._on_profile_custom_count_toggled,
-        )
-        self._profile_custom_count_row.append(self._profile_custom_count_toggle)
-        self._profile_lifetime_count_spin = Gtk.SpinButton()
-        self._profile_lifetime_count_spin.set_size_request(92, -1)
-        self._profile_lifetime_count_spin.set_adjustment(
-            Gtk.Adjustment(
-                value=self._profile_lifetime_action_count,
-                lower=1,
-                upper=999,
-                step_increment=1,
-            )
-        )
-        self._profile_lifetime_count_spin.connect(
-            "value-changed",
-            self._on_profile_lifetime_count_changed,
-        )
-        self._profile_custom_count_row.append(self._profile_lifetime_count_spin)
-        self._profile_lifetime_custom_box.append(self._profile_custom_count_row)
-
-        self._profile_custom_timeout_row = Gtk.Box(
-            orientation=Gtk.Orientation.HORIZONTAL,
-            spacing=8,
-        )
-        self._profile_custom_timeout_row.set_halign(Gtk.Align.START)
-        self._profile_custom_timeout_toggle = Gtk.ToggleButton(label="Timeout")
-        self._profile_custom_timeout_toggle.set_size_request(140, -1)
-        self._profile_custom_timeout_toggle.set_active(self._profile_custom_timeout)
-        self._profile_custom_timeout_toggle.connect(
-            "toggled",
-            self._on_profile_custom_timeout_toggled,
-        )
-        self._profile_custom_timeout_row.append(self._profile_custom_timeout_toggle)
-        self._profile_custom_timeout_spin = Gtk.SpinButton()
-        self._profile_custom_timeout_spin.set_size_request(92, -1)
-        self._profile_custom_timeout_spin.set_adjustment(self._profile_lifetime_timeout_adjustment)
-        self._profile_custom_timeout_spin.connect(
-            "value-changed",
-            self._on_profile_lifetime_timeout_changed,
-        )
-        self._profile_custom_timeout_row.append(self._profile_custom_timeout_spin)
-        self._profile_custom_timeout_unit = Gtk.Label(label="ms")
-        self._profile_custom_timeout_unit.add_css_class("dim-label")
-        self._profile_custom_timeout_row.append(self._profile_custom_timeout_unit)
-        self._profile_lifetime_custom_box.append(self._profile_custom_timeout_row)
-
-        self._profile_custom_trigger_toggle = Gtk.ToggleButton(label="Trigger end")
-        self._profile_custom_trigger_toggle.set_halign(Gtk.Align.START)
-        self._profile_custom_trigger_toggle.set_size_request(140, -1)
-        self._profile_custom_trigger_toggle.set_active(self._profile_custom_trigger_end)
-        self._profile_custom_trigger_toggle.connect(
-            "toggled",
-            self._on_profile_custom_trigger_toggled,
-        )
-        self._profile_lifetime_custom_box.append(self._profile_custom_trigger_toggle)
-
-        box.append(self._profile_lifetime_custom_box)
-
-        self._profile_lifetime_notice_label = Gtk.Label(
-            label="Disable this profile first to use it as a temporary layer."
-        )
-        self._profile_lifetime_notice_label.add_css_class("dim-label")
-        self._profile_lifetime_notice_label.set_halign(Gtk.Align.START)
-        self._profile_lifetime_notice_label.set_wrap(True)
-        box.append(self._profile_lifetime_notice_label)
-
-        self._update_profile_lifetime_visibility()
-        return box
-
-    def _on_profile_lifetime_preset_changed(self, dropdown, _pspec) -> None:
-        if self._profile_lifetime_selection_updating:
-            return
-        if self._profile_lifetime_blocked():
-            self._sync_profile_lifetime_dropdown("until_changed")
-            self._update_profile_lifetime_visibility()
-            self._update_profile_hint()
-            return
-        idx = int(dropdown.get_selected())
-        presets = self._profile_lifetime_presets()
-        self._profile_lifetime_preset = (
-            presets[idx][0] if 0 <= idx < len(presets) else "until_changed"
-        )
-        self._update_profile_lifetime_visibility()
-        self._update_profile_hint()
-
-    def _on_profile_lifetime_count_changed(self, spin: Gtk.SpinButton) -> None:
-        self._profile_lifetime_action_count = int(spin.get_value())
-
-    def _on_profile_lifetime_timeout_changed(self, spin: Gtk.SpinButton) -> None:
-        self._profile_lifetime_timeout_ms = int(spin.get_value())
-
-    def _on_profile_custom_trigger_toggled(self, check: Gtk.ToggleButton) -> None:
-        self._profile_custom_trigger_end = bool(check.get_active())
-
-    def _on_profile_custom_count_toggled(self, check: Gtk.ToggleButton) -> None:
-        self._profile_custom_action_count = bool(check.get_active())
-        self._update_profile_lifetime_visibility()
-
-    def _on_profile_custom_timeout_toggled(self, check: Gtk.ToggleButton) -> None:
-        self._profile_custom_timeout = bool(check.get_active())
-        self._update_profile_lifetime_visibility()
-
-    def _profile_lifetime_presets(self) -> tuple[tuple[str, str], ...]:
-        if self._selected_profile_action == "toggle":
-            return _PROFILE_LIFETIME_PRESETS_TOGGLE
-        return _PROFILE_LIFETIME_PRESETS_ENABLE
-
-    def _coerce_profile_lifetime_preset(self, preset: str) -> str:
-        keys = {key for key, _label in self._profile_lifetime_presets()}
-        return preset if preset in keys else "until_changed"
-
-    def _profile_lifetime_preset_index(self, preset: str) -> int:
-        preset = self._coerce_profile_lifetime_preset(preset)
-        for idx, (key, _label) in enumerate(self._profile_lifetime_presets()):
-            if key == preset:
-                return idx
-        return 0
-
-    def _sync_profile_lifetime_model(self) -> None:
-        if not hasattr(self, "_profile_lifetime_model"):
-            return
-        presets = self._profile_lifetime_presets()
-        keys = [key for key, _label in presets]
-        if self._profile_lifetime_model_keys == keys:
-            return
-        self._profile_lifetime_selection_updating = True
-        try:
-            while self._profile_lifetime_model.get_n_items() > 0:
-                self._profile_lifetime_model.remove(0)
-            for _key, label in presets:
-                self._profile_lifetime_model.append(label)
-            self._profile_lifetime_model_keys = keys
-        finally:
-            self._profile_lifetime_selection_updating = False
-
-    def _sync_profile_lifetime_dropdown(self, preset: str) -> None:
-        if not hasattr(self, "_profile_lifetime_dropdown"):
-            return
-        self._sync_profile_lifetime_model()
-        preset = self._coerce_profile_lifetime_preset(preset)
-        idx = self._profile_lifetime_preset_index(preset)
-        if int(self._profile_lifetime_dropdown.get_selected()) == idx:
-            return
-        self._profile_lifetime_selection_updating = True
-        try:
-            self._profile_lifetime_dropdown.set_selected(idx)
-        finally:
-            self._profile_lifetime_selection_updating = False
-
-    def _selected_profile_entry(self) -> dict | None:
-        for profile in self._profile_entries:
-            if str(profile.get("name", "") or "") == self._selected_profile_name:
-                return profile
-        return None
-
-    def _selected_profile_enabled(self) -> bool:
-        profile = self._selected_profile_entry()
-        return bool(profile and bool(profile.get("enabled", True)))
-
-    def _profile_lifetime_available(self) -> bool:
-        return self._selected_profile_action in {"enable", "toggle"}
-
-    def _profile_lifetime_blocked(self) -> bool:
-        return self._profile_lifetime_available() and self._selected_profile_enabled()
-
-    def _update_profile_lifetime_visibility(self) -> None:
-        if not hasattr(self, "_profile_lifetime_box"):
-            return
-        lifetime_available = self._profile_lifetime_available()
-        blocked = self._profile_lifetime_blocked()
-        preset = (
-            "until_changed"
-            if blocked
-            else self._coerce_profile_lifetime_preset(self._profile_lifetime_preset)
-        )
-        if not blocked:
-            self._profile_lifetime_preset = preset
-        is_custom = preset == "custom"
-        is_one_shot = preset == "after_one_action"
-        allow_trigger_end = self._selected_profile_action == "enable"
-        if not allow_trigger_end and self._profile_custom_trigger_end:
-            self._profile_custom_trigger_end = False
-            self._profile_custom_trigger_toggle.set_active(False)
-        self._sync_profile_lifetime_dropdown(preset)
-        self._profile_lifetime_box.set_visible(lifetime_available)
-        self._profile_lifetime_title.set_label(
-            "When toggled on" if self._selected_profile_action == "toggle" else "Activation"
-        )
-        self._profile_lifetime_dropdown.set_sensitive(not blocked)
-        self._profile_lifetime_timeout_row.set_visible(False)
-        self._profile_lifetime_custom_box.set_visible(is_custom or is_one_shot)
-        self._profile_custom_count_row.set_visible(is_custom)
-        self._profile_custom_timeout_row.set_visible(is_custom or is_one_shot)
-        self._profile_custom_trigger_toggle.set_visible(is_custom and allow_trigger_end)
-        self._profile_lifetime_count_spin.set_sensitive(self._profile_custom_action_count)
-        self._profile_custom_timeout_spin.set_visible(self._profile_custom_timeout)
-        self._profile_custom_timeout_spin.set_sensitive(self._profile_custom_timeout)
-        self._profile_custom_timeout_unit.set_visible(self._profile_custom_timeout)
-        self._profile_lifetime_notice_label.set_visible(blocked)
-
-    def _populate_profile_names(self) -> None:
-        profile_name_items: list[str] = []
-        profile_name_labels: list[str] = []
-        for profile in self._profile_entries:
-            name = str(profile.get("name", "") or "")
-            if not name:
-                continue
-            profile_name_items.append(name)
-            enabled = bool(profile.get("enabled", True))
-            marker = "" if enabled else " [disabled]"
-            profile_name_labels.append(f"{name}{marker}")
-
-        self._profile_name_populating = True
-        try:
-            self._profile_name_items = profile_name_items
-            while self._profile_name_model.get_n_items() > 0:
-                self._profile_name_model.remove(0)
-            for label in profile_name_labels:
-                self._profile_name_model.append(label)
-        finally:
-            self._profile_name_populating = False
-
-        if not self._profile_name_items:
-            self._selected_profile_name = ""
-            self._update_profile_hint()
-            self._on_tab_changed(self.stack, None)
-            return
-
-        selected_index = 0
-        if self._selected_profile_name in self._profile_name_items:
-            selected_index = self._profile_name_items.index(self._selected_profile_name)
-        else:
-            self._selected_profile_name = self._profile_name_items[0]
-
-        self._profile_name_dropdown.set_selected(selected_index)
-        self._on_profile_name_changed(self._profile_name_dropdown, None)
-
-    def _on_profile_name_changed(self, dropdown, _pspec) -> None:
-        if self._profile_name_populating or not self._profile_name_items:
-            return
-        idx = int(dropdown.get_selected())
-        if idx < 0 or idx >= len(self._profile_name_items):
-            self._selected_profile_name = ""
-        else:
-            self._selected_profile_name = self._profile_name_items[idx]
-        self._update_profile_lifetime_visibility()
-        self._update_profile_hint()
-        self._on_tab_changed(self.stack, None)
-
-    def _update_profile_hint(self) -> None:
-        if not self._selected_profile_name:
-            self._profile_hint_label.set_label("Select a profile to map this action.")
-            return
-
-        verb = {
-            "toggle": "Toggle",
-            "enable": "Enable",
-            "disable": "Disable",
-        }.get(self._selected_profile_action, "Toggle")
-        self._profile_hint_label.set_label(f"{verb} profile '{self._selected_profile_name}'.")
-
-    def _load_superkey_list(self) -> None:
-        manager = SuperkeyManager()
-        configs = manager.get_all_superkeys()
-        self._superkey_names = manager.list_superkeys()
-        self._superkey_list = [
-            config for name in self._superkey_names if (config := configs.get(name)) is not None
-        ]
-        self._populate_superkey_listbox()
-
-    def _populate_superkey_listbox(self) -> None:
-        while self._superkey_listbox.get_first_child():
-            self._superkey_listbox.remove(self._superkey_listbox.get_first_child())
-
-        if not self._superkey_list:
-            self._selected_superkey = None
-            row = Gtk.ListBoxRow()
-            row.set_selectable(False)
-            lbl = Gtk.Label(label="No super keys saved yet")
-            lbl.add_css_class("dim-label")
-            lbl.set_margin_top(12)
-            lbl.set_margin_bottom(12)
-            row.set_child(lbl)
-            self._superkey_listbox.append(row)
-            return
-
-        selected_row: Gtk.ListBoxRow | None = None
-        for config in self._superkey_list:
-            row = Gtk.ListBoxRow()
-            row._superkey_name = config.name
-            right_click = Gtk.GestureClick()
-            right_click.set_button(3)
-            right_click.connect("pressed", self._on_superkey_row_right_pressed, config.name)
-            row.add_controller(right_click)
-
-            row_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=12)
-            row_box.set_margin_top(8)
-            row_box.set_margin_bottom(8)
-            row_box.set_margin_start(12)
-            row_box.set_margin_end(12)
-
-            name_label = Gtk.Label(label=config.name)
-            name_label.set_halign(Gtk.Align.START)
-            name_label.set_hexpand(True)
-            row_box.append(name_label)
-
-            info_label = Gtk.Label(label=self._describe_superkey_row(config))
-            info_label.add_css_class("dim-label")
-            info_label.add_css_class("caption")
-            row_box.append(info_label)
-
-            row.set_child(row_box)
-            self._superkey_listbox.append(row)
-
-            if self._selected_superkey and config.name == self._selected_superkey:
-                selected_row = row
-
-        if selected_row is not None:
-            self._superkey_listbox.select_row(selected_row)
-        elif self._selected_superkey:
-            self._selected_superkey = None
-
-    def _describe_superkey_row(self, config: SuperkeyConfig) -> str:
-        if config.mode.value == "overload":
-            count = (
-                len(config.overload_actions)
-                + len(config.overload_down_actions)
-                + len(config.overload_up_actions)
-            )
-            noun = "action" if count == 1 else "actions"
-            suffix = (
-                " · down/up"
-                if config.overload_down_actions or config.overload_up_actions
-                else ""
-            )
-            return f"Overload{suffix} · {count} {noun}"
-
-        slots = sum(
-            1
-            for actions in (
-                config.tap_actions,
-                config.double_tap_actions,
-                config.hold_actions,
-                config.tap_hold_actions,
-            )
-            if actions
-        )
-        noun = "slot" if slots == 1 else "slots"
-        return f"Pattern · {slots} {noun}"
-
-    def _on_superkey_row_right_pressed(
-        self,
-        gesture: Gtk.GestureClick,
-        _n_press: int,
-        _x: float,
-        _y: float,
-        name: str,
-    ) -> None:
-        gesture.set_state(Gtk.EventSequenceState.CLAIMED)
-        self._open_superkey_manager(name)
-
-    def _on_superkey_row_selected(self, listbox, row) -> None:
-        if row and hasattr(row, "_superkey_name"):
-            self._selected_superkey = row._superkey_name
-        else:
-            self._selected_superkey = None
-        if self.stack.get_visible_child_name() == "superkey":
-            self.map_btn.set_sensitive(self._selected_superkey is not None)
-
-    def _load_analog_control_list(self) -> None:
-        manager = AnalogControlManager()
-        configs = manager.get_all_analog_controls()
-        self._analog_control_names = manager.list_analog_controls()
-        self._analog_control_list = [
-            config
-            for name in self._analog_control_names
-            if (config := configs.get(name)) is not None
-            and (
-                self._analog_input_type is None
-                or config.input_type == self._analog_input_type
-            )
-        ]
-        self._populate_analog_control_listbox()
-
-    def _populate_analog_control_listbox(self) -> None:
-        selected_names = set(self._selected_analog_controls)
-        while self._analog_control_listbox.get_first_child():
-            self._analog_control_listbox.remove(self._analog_control_listbox.get_first_child())
-
-        if not self._analog_control_list:
-            self._selected_analog_control = None
-            self._selected_analog_controls = []
-            row = Gtk.ListBoxRow()
-            row.set_selectable(False)
-            label = "No analog controls saved yet"
-            if self._analog_input_type == "axis":
-                label = "No axis controls saved yet"
-            elif self._analog_input_type == "stick":
-                label = "No stick controls saved yet"
-            lbl = Gtk.Label(label=label)
-            lbl.add_css_class("dim-label")
-            lbl.set_margin_top(12)
-            lbl.set_margin_bottom(12)
-            row.set_child(lbl)
-            self._analog_control_listbox.append(row)
-            return
-
-        for config in self._analog_control_list:
-            row = Gtk.ListBoxRow()
-            row._analog_control_name = config.name
-            click = Gtk.GestureClick()
-            click.set_button(1)
-            click.set_propagation_phase(Gtk.PropagationPhase.CAPTURE)
-            click.connect("pressed", self._on_analog_control_row_pressed, row)
-            row.add_controller(click)
-
-            right_click = Gtk.GestureClick()
-            right_click.set_button(3)
-            right_click.connect(
-                "pressed", self._on_analog_control_row_right_pressed, config.name
-            )
-            row.add_controller(right_click)
-
-            row_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=12)
-            row_box.set_margin_top(8)
-            row_box.set_margin_bottom(8)
-            row_box.set_margin_start(12)
-            row_box.set_margin_end(12)
-
-            name_label = Gtk.Label(label=config.name)
-            name_label.set_halign(Gtk.Align.START)
-            name_label.set_hexpand(True)
-            row_box.append(name_label)
-
-            info_label = Gtk.Label(label=self._describe_analog_control_row(config))
-            info_label.add_css_class("dim-label")
-            info_label.add_css_class("caption")
-            row_box.append(info_label)
-
-            row.set_child(row_box)
-            self._analog_control_listbox.append(row)
-
-            if config.name in selected_names:
-                self._analog_control_listbox.select_row(row)
-        self._sync_selected_analog_controls()
-
-    def _describe_analog_control_row(self, config: AnalogControlConfig) -> str:
-        parts: list[str] = ["Axis" if config.input_type == "axis" else "Stick"]
-        if config.mouse_motion.enabled:
-            parts.append("Mouse")
-        if config.gamepad_output.enabled:
-            parts.append("Gamepad output")
-        if config.thresholds:
-            count = len(config.thresholds)
-            parts.append(f"{count} range{'s' if count != 1 else ''}")
-        return " · ".join(parts)
-
-    def _on_analog_control_row_right_pressed(
-        self,
-        gesture: Gtk.GestureClick,
-        _n_press: int,
-        _x: float,
-        _y: float,
-        name: str,
-    ) -> None:
-        gesture.set_state(Gtk.EventSequenceState.CLAIMED)
-        self._open_analog_control_manager(name)
-
-    def _on_analog_control_row_pressed(
-        self,
-        gesture: Gtk.GestureClick,
-        _n_press: int,
-        _x: float,
-        _y: float,
-        row: Gtk.ListBoxRow,
-    ) -> None:
-        if row.is_selected():
-            gesture.set_state(Gtk.EventSequenceState.CLAIMED)
-            self._unselect_analog_control_row(row)
-
-    def _unselect_analog_control_row(self, row: Gtk.ListBoxRow) -> bool:
-        if row.is_selected():
-            self._analog_control_listbox.unselect_row(row)
-            self._sync_selected_analog_controls()
-            if self.stack.get_visible_child_name() == "analog_control":
-                self.map_btn.set_sensitive(bool(self._selected_analog_controls))
-        return False
-
-    def _on_analog_control_row_selected(self, listbox, row) -> None:
-        self._sync_selected_analog_controls()
-        if self.stack.get_visible_child_name() == "analog_control":
-            self.map_btn.set_sensitive(bool(self._selected_analog_controls))
-
-    def _sync_selected_analog_controls(self) -> None:
-        selected: list[str] = []
-        for row in self._analog_control_listbox.get_selected_rows():
-            name = getattr(row, "_analog_control_name", None)
-            if isinstance(name, str):
-                selected.append(name)
-        self._selected_analog_controls = selected
-        self._selected_analog_control = selected[0] if selected else None
-
-    def _load_macro_list(self) -> bool:
-        session_request_async({"command": "list_macros"}, self._on_macro_list_loaded)
-        return False
-
-    def _on_macro_list_loaded(self, result: dict | None) -> bool:
-        self._macro_list = (result or {}).get("macros", [])
-        self._populate_macro_listbox()
-        return False
-
-    def _populate_macro_listbox(self) -> None:
-        while self._macro_listbox.get_first_child():
-            self._macro_listbox.remove(self._macro_listbox.get_first_child())
-
-        if not self._macro_list:
-            row = Gtk.ListBoxRow()
-            row.set_selectable(False)
-            lbl = Gtk.Label(label="No macros saved yet")
-            lbl.add_css_class("dim-label")
-            lbl.set_margin_top(12)
-            lbl.set_margin_bottom(12)
-            row.set_child(lbl)
-            self._macro_listbox.append(row)
-            return
-
-        for macro in self._macro_list:
-            row = Gtk.ListBoxRow()
-            row._macro_name = macro["name"]
-            row._search_text = macro_search_text(macro)
-            right_click = Gtk.GestureClick()
-            right_click.set_button(3)
-            right_click.connect("pressed", self._on_macro_row_right_pressed, macro["name"])
-            row.add_controller(right_click)
-
-            row_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=12)
-            row_box.set_margin_top(8)
-            row_box.set_margin_bottom(8)
-            row_box.set_margin_start(12)
-            row_box.set_margin_end(12)
-
-            name_label = Gtk.Label(label=macro["name"])
-            name_label.set_halign(Gtk.Align.START)
-            name_label.set_hexpand(True)
-            row_box.append(name_label)
-
-            duration_s = macro.get("duration_us", 0) / 1_000_000
-            device_types = ", ".join(macro.get("device_types", []))
-            info_label = Gtk.Label(label=f"{duration_s:.1f}s · {device_types}")
-            info_label.add_css_class("dim-label")
-            info_label.add_css_class("caption")
-            row_box.append(info_label)
-
-            row.set_child(row_box)
-            self._macro_listbox.append(row)
-        self._macro_listbox.invalidate_filter()
-
-        if self._selected_macro:
-            for i, macro in enumerate(self._macro_list):
-                if macro["name"] == self._selected_macro:
-                    row = self._macro_listbox.get_row_at_index(i)
-                    if row and fuzzy_query_matches(
-                        self._macro_search_entry.get_text(),
-                        getattr(row, "_search_text", ""),
-                    ):
-                        self._macro_listbox.select_row(row)
-                    else:
-                        self._clear_macro_selection()
-                    break
-            else:
-                self._clear_macro_selection()
-        if not self._selected_macro:
-            self._clear_macro_selection()
-
-    def _after_macro_search_filter_changed(self) -> None:
-        selected_row = self._macro_listbox.get_selected_row()
-        if selected_row is None:
-            self._clear_macro_selection()
-            return
-        if fuzzy_query_matches(
-            self._macro_search_entry.get_text(),
-            getattr(selected_row, "_search_text", ""),
-        ):
-            return
-        self._macro_listbox.unselect_row(selected_row)
-        self._clear_macro_selection()
-
-    def _clear_macro_selection(self) -> None:
-        self._selected_macro = None
-        self._macro_options_box.set_visible(False)
-        if self.stack.get_visible_child_name() == "macro":
-            self.map_btn.set_sensitive(False)
-
-    def _on_macro_row_selected(self, listbox, row) -> None:
-        if row and hasattr(row, "_macro_name"):
-            self._selected_macro = row._macro_name
-            self._macro_options_box.set_visible(self._allow_macro_options)
-        else:
-            self._clear_macro_selection()
-        self.map_btn.set_sensitive(self._selected_macro is not None)
-
-    def _on_macro_row_right_pressed(
-        self,
-        gesture: Gtk.GestureClick,
-        _n_press: int,
-        _x: float,
-        _y: float,
-        name: str,
-    ) -> None:
-        gesture.set_state(Gtk.EventSequenceState.CLAIMED)
-        self._open_macro_editor(name)
-
-    def _open_macro_editor(self, name: str) -> None:
-        from keymasq.gui.widgets.macro_editor_dialog import MacroEditorDialog
-
-        root = self.get_root()
-        parent = root if root is not None else self._parent
-        dialog = MacroEditorDialog(parent, name)
-        dialog.connect("closed", self._on_macro_editor_closed)
-        dialog.present(parent)
-
-    def _on_macro_editor_closed(self, _dialog: Adw.Dialog) -> None:
-        self._load_macro_list()
-
-    def _on_macro_movement_toggled(self, check) -> None:
-        self._macro_replay_movement = check.get_active()
-
-    def _on_macro_clicks_toggled(self, check) -> None:
-        self._macro_replay_clicks = check.get_active()
-
-    def _on_macro_speed_changed(self, spin) -> None:
-        self._macro_speed = spin.get_value()
-
-    def _on_macro_special_action_clicked(self, _btn, action_name: str) -> None:
-        if action_name == "cancel_macro_playback":
-            self._warn_and_clear_unsupported_rapidfire(ActionType.CANCEL_MACRO_PLAYBACK)
-            action = MappingAction(action_type=ActionType.CANCEL_MACRO_PLAYBACK)
-        else:
-            return
-        self.emit("key-selected", action)
-        self.close()
-
-    def _on_macro_recording_slot_clicked(self, _btn, slot: int) -> None:
-        self._warn_and_clear_unsupported_rapidfire(ActionType.START_MACRO_RECORDING)
-        action = MappingAction(
-            action_type=ActionType.START_MACRO_RECORDING,
-            macro_recording_slot=slot,
-        )
-        self.emit("key-selected", action)
-        self.close()
-
-    def _on_macro_play_slot_clicked(self, _btn, slot: int) -> None:
-        self._warn_and_clear_unsupported_rapidfire(ActionType.PLAY_MACRO_SLOT)
-        action = MappingAction(
-            action_type=ActionType.PLAY_MACRO_SLOT,
-            macro_recording_slot=slot,
-        )
-        self.emit("key-selected", action)
-        self.close()
-
-    def _on_map_clicked(self, btn) -> None:
-        child_name = self.stack.get_visible_child_name()
-        if child_name == "superkey":
-            self._on_superkey_map_clicked(btn)
-        elif child_name == "analog_control":
-            self._on_analog_control_map_clicked(btn)
-        elif child_name == "macro":
-            self._on_macro_map_clicked(btn)
-        elif child_name == "profile":
-            self._on_profile_map_clicked(btn)
-
-    def _on_superkey_map_clicked(self, btn) -> None:
-        if not self._selected_superkey:
-            return
-        self._warn_and_clear_unsupported_rapidfire(ActionType.SUPERKEY)
-        action = MappingAction(
-            action_type=ActionType.SUPERKEY,
-            superkey_name=self._selected_superkey,
-        )
-        self.emit("key-selected", action)
-        self.close()
-
-    def _on_analog_control_map_clicked(self, btn) -> None:
-        if not self._selected_analog_controls:
-            return
-        action = MappingAction(
-            action_type=ActionType.ANALOG_CONTROL,
-            analog_control_names=list(self._selected_analog_controls),
-        )
-        self.emit("key-selected", action)
-        self.close()
-
-    def _on_macro_map_clicked(self, btn) -> None:
-        if not self._selected_macro:
-            return
-        self._warn_and_clear_unsupported_rapidfire(ActionType.MACRO)
-        action = MappingAction(
-            action_type=ActionType.MACRO,
-            macro_name=self._selected_macro,
-            macro_replay_mouse_movement=self._macro_replay_movement,
-            macro_replay_mouse_clicks=self._macro_replay_clicks,
-            macro_speed=self._macro_speed,
-        )
-        self.emit("key-selected", action)
-        self.close()
-
-    def _on_profile_map_clicked(self, btn) -> None:
-        if not self._selected_profile_name:
-            return
-
-        action_type = ActionType.PROFILE_TOGGLE
-        if self._selected_profile_action == "enable":
-            action_type = ActionType.PROFILE_ENABLE
-        elif self._selected_profile_action == "disable":
-            action_type = ActionType.PROFILE_DISABLE
-
-        self._warn_and_clear_unsupported_rapidfire(action_type)
-        action = MappingAction(
-            action_type=action_type,
-            profile_name=self._selected_profile_name,
-            profile_deactivation=self._profile_deactivation_policy(action_type),
-        )
-        self.emit("key-selected", action)
-        self.close()
-
-    def _profile_deactivation_policy(
-        self,
-        action_type: ActionType,
-    ) -> ProfileDeactivationPolicy | None:
-        if action_type not in (ActionType.PROFILE_ENABLE, ActionType.PROFILE_TOGGLE):
-            return None
-        if self._profile_lifetime_blocked():
-            return None
-        preset = self._coerce_profile_lifetime_preset(self._profile_lifetime_preset)
-        if preset == "while_trigger_active":
-            if action_type == ActionType.PROFILE_ENABLE:
-                return ProfileDeactivationPolicy(on_trigger_end=True)
-            return None
-        if preset == "after_one_action":
-            timeout_ms = (
-                max(1, int(self._profile_lifetime_timeout_ms))
-                if self._profile_custom_timeout
-                else None
-            )
-            return ProfileDeactivationPolicy(after_actions=1, timeout_ms=timeout_ms)
-        if preset != "custom":
-            return None
-
-        after_actions = (
-            max(1, int(self._profile_lifetime_action_count))
-            if self._profile_custom_action_count
-            else None
-        )
-        timeout_ms = (
-            max(1, int(self._profile_lifetime_timeout_ms))
-            if self._profile_custom_timeout
-            else None
-        )
-        policy = ProfileDeactivationPolicy(
-            on_trigger_end=(
-                bool(self._profile_custom_trigger_end)
-                if action_type == ActionType.PROFILE_ENABLE
-                else False
-            ),
-            after_actions=after_actions,
-            timeout_ms=timeout_ms,
-        )
-        return policy if policy.has_condition else None
-
-    def _set_initial_tab(self):
-        if self._source_type == "analog":
-            self._set_initial_analog_tab()
-            return
-        if not self._current_action:
-            return
-        compositor_tab = compositor_action_tab_name(
-            self._current_action,
-            self._compositor_action_status,
-        )
-        if compositor_tab not in self._compositor_action_page_ids:
-            compositor_tab = None
-        tab_map = {
-            ActionType.PASSTHROUGH: "special",
-            ActionType.SUPPRESS: "special",
-            ActionType.REPEAT: "special",
-            ActionType.SUPERKEY: "superkey",
-            ActionType.ANALOG_CONTROL: "analog_control",
-            ActionType.START_MACRO_RECORDING: "macro",
-            ActionType.STOP_MACRO_RECORDING: "macro",
-            ActionType.PLAY_MACRO_SLOT: "macro",
-            ActionType.CANCEL_MACRO_PLAYBACK: "macro",
-            ActionType.EMERGENCY_RESET: "macro",
-            ActionType.EXEC: "special",
-            ActionType.MPRIS: "media",
-            ActionType.KEYBOARD: (
-                "media" if self._current_action.target in MEDIA_KEY_TARGETS else "keyboard"
-            ),
-            ActionType.MOUSE: "mouse",
-            ActionType.MOUSE_MOVE_REL: "mouse",
-            ActionType.MOUSE_MOVE_ABS: "mouse",
-            ActionType.GAMEPAD: "gamepad",
-            ActionType.GAMEPAD_AXIS: "gamepad",
-            ActionType.MACRO: "macro",
-            ActionType.PROFILE_ENABLE: "profile",
-            ActionType.PROFILE_DISABLE: "profile",
-            ActionType.PROFILE_TOGGLE: "profile",
-        }
-        name = compositor_tab or tab_map.get(self._current_action.action_type)
-        if name == "superkey" and not self._allow_superkey:
-            return
-        if name:
-            self.stack.set_visible_child_name(name)
-
-    def _resolve_compositor_action_status(
-        self,
-        compositor_action_status: dict[str, object] | None,
-    ) -> dict[str, bool | str | None]:
-        resolved: dict[str, bool | str | None] = {
-            "compositor_id": None,
-            "listener_name": None,
-            "compositor_dispatch_available": False,
-        }
-        if isinstance(compositor_action_status, dict):
-            for key in resolved:
-                value = compositor_action_status.get(key)
-                if isinstance(value, (bool, str)) or value is None:
-                    resolved[key] = value
-            return resolved
-
-        root = self._parent.get_root() if hasattr(self._parent, "get_root") else None
-        get_status = getattr(root, "get_compositor_action_status", None)
-        if callable(get_status):
-            status = get_status()
-            if isinstance(status, dict):
-                for key in resolved:
-                    value = status.get(key)
-                    if isinstance(value, (bool, str)) or value is None:
-                        resolved[key] = value
-        return resolved
-
-    def _describe_current_action(self) -> str:
-        return describe_mapping_action_verbose(
-            self._current_action,
-            keyboard_label=lambda value: EVDEV_TO_KEY.get(value, value),
-            gamepad_label=lambda value: EVDEV_TO_GAMEPAD.get(value, value),
-        )
-
-    def _on_f_dropdown_changed(self, dropdown, pspec, btn: Gtk.Button):
-        idx = dropdown.get_selected()
-        f_key = F_EXTRA[idx]
-        btn.set_label(f"Map {f_key}")
-
-
-class SuperkeyActionDialog(Adw.Dialog, _GamepadAxisControlsMixin):
-    __gsignals__ = {
-        "action-selected": (GObject.SignalFlags.RUN_FIRST, None, (object,)),
-    }
-
-    def __init__(
-        self,
-        parent: Gtk.Widget,
-        action_type: str,
-        current_action: SuperkeyAction | None = None,
-    ):
-        self._action_type = action_type
-        self._current_action = current_action
-        self._parent = parent
-
-        title = f"Configure {action_type.replace('_', ' ').title()} Action"
-        super().__init__(title=title, content_width=540, content_height=520)
-
-        self._rapidfire_enabled = False
-        self._rapidfire_hold = 20
-        self._rapidfire_wait = 20
-        self._superkey_macro_list: list[dict] = []
-        self._superkey_selected_macro: str | None = None
-        self._selected_gamepad_output_id: str | None = (
-            current_action.output_id
-            if current_action
-            and current_action.action_type in (ActionType.GAMEPAD, ActionType.GAMEPAD_AXIS)
-            else None
-        )
-        self._gamepad_output_ids: list[str | None] = []
-        self._gamepad_output_warning_label: Gtk.Label | None = None
-
-        if current_action:
-            self._rapidfire_enabled = current_action.rapidfire_enabled
-            self._rapidfire_hold = current_action.rapidfire_hold_ms
-            self._rapidfire_wait = current_action.rapidfire_wait_ms
-            if current_action.action_type == ActionType.MACRO:
-                self._superkey_selected_macro = current_action.macro_name
-
-        self._build_ui()
-
-    def _build_ui(self):
-        main_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=0)
-        main_box.set_margin_top(12)
-        main_box.set_margin_bottom(12)
-        main_box.set_margin_start(12)
-        main_box.set_margin_end(12)
-
-        self.options_box = self._build_options_box()
-
-        self.stack = Gtk.Stack()
-        self.stack.set_vexpand(True)
-        self.stack.connect("notify::visible-child", self._on_tab_changed)
-        self.stack.add_titled(self._build_keyboard_tab(), "keyboard", "Keyboard")
-        self.stack.add_titled(self._build_navigation_tab(), "navigation", "Navigation")
-        self.stack.add_titled(self._build_media_tab(), "media", "Media")
-        self.stack.add_titled(self._build_mouse_tab(), "mouse", "Mouse")
-        self.stack.add_titled(self._build_gamepad_tab(), "gamepad", "Gamepad")
-        self.stack.add_titled(self._build_macro_tab(), "macro", "Macro")
-        self.stack.add_titled(self._build_exec_tab(), "exec", "Command")
-
-        self._set_initial_tab()
-
-        frame = Gtk.Frame()
-        frame.set_vexpand(True)
-
-        inner = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=0)
-
-        title_label = Gtk.Label(label=self.get_title())
-        title_label.add_css_class("title-3")
-        title_label.set_halign(Gtk.Align.CENTER)
-        title_label.set_margin_bottom(12)
-        inner.append(title_label)
-
-        inner.append(Gtk.Separator())
-
-        switcher = Gtk.StackSwitcher()
-        switcher.set_stack(self.stack)
-        switcher.set_halign(Gtk.Align.CENTER)
-        switcher.set_margin_top(8)
-        switcher.set_margin_bottom(8)
-        inner.append(switcher)
-
-        inner.append(Gtk.Separator())
-        inner.append(self.stack)
-        inner.append(Gtk.Separator())
-
-        options_pad = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=0)
-        options_pad.set_margin_top(8)
-        options_pad.set_margin_bottom(8)
-        options_pad.set_margin_start(12)
-        options_pad.set_margin_end(12)
-        options_pad.append(self.options_box)
-        inner.append(options_pad)
-
-        inner.append(Gtk.Separator())
-
-        btn_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        btn_box.set_hexpand(True)
-        btn_box.set_margin_top(8)
-        btn_box.set_margin_bottom(8)
-        btn_box.set_margin_start(12)
-        btn_box.set_margin_end(12)
-
-        self.actions_docs_btn = _create_actions_docs_button()
-        self.actions_docs_btn.connect("clicked", self._on_actions_docs_clicked)
-        btn_box.append(self.actions_docs_btn)
-
-        footer_spacer = Gtk.Box()
-        footer_spacer.set_hexpand(True)
-        btn_box.append(footer_spacer)
-
-        clear_btn = Gtk.Button(label="Clear")
-        clear_btn.connect("clicked", self._on_clear_clicked)
-        btn_box.append(clear_btn)
-
-        cancel_btn = Gtk.Button(label="Cancel")
-        cancel_btn.connect("clicked", self._on_cancel_clicked)
-        btn_box.append(cancel_btn)
-
-        inner.append(btn_box)
-
-        frame.set_child(inner)
-        main_box.append(frame)
-        self.set_child(main_box)
-
-        GLib.idle_add(self._load_superkey_macro_list)
-
-    def _on_cancel_clicked(self, _button: Gtk.Button) -> None:
-        self.close()
-
-    def _build_options_box(self) -> Gtk.Widget:
-        box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=8)
-
-        is_hold = self._action_type in ("hold", "tap_hold")
-
-        if is_hold:
-            row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-            row.set_halign(Gtk.Align.START)
-
-            self.rapidfire_check = Gtk.CheckButton(label="Rapidfire")
-            self.rapidfire_check.set_active(self._rapidfire_enabled)
-            self.rapidfire_check.set_tooltip_text("Repeatedly send while held")
-            self.rapidfire_check.connect("toggled", self._on_rapidfire_toggled)
-            row.append(self.rapidfire_check)
-
-            self.hold_label = Gtk.Label(label="Hold:")
-            row.append(self.hold_label)
-
-            self.hold_spin = Gtk.SpinButton()
-            hold_adj = Gtk.Adjustment(
-                value=self._rapidfire_hold,
-                lower=MIN_RAPIDFIRE_HOLD_MS,
-                upper=1000,
-                step_increment=1,
-            )
-            self.hold_spin.set_adjustment(hold_adj)
-            row.append(self.hold_spin)
-
-            self.hold_ms_label = Gtk.Label(label="ms")
-            row.append(self.hold_ms_label)
-
-            self.wait_label = Gtk.Label(label="Wait:")
-            row.append(self.wait_label)
-
-            self.wait_spin = Gtk.SpinButton()
-            wait_adj = Gtk.Adjustment(
-                value=self._rapidfire_wait,
-                lower=MIN_RAPIDFIRE_WAIT_MS,
-                upper=1000,
-                step_increment=1,
-            )
-            self.wait_spin.set_adjustment(wait_adj)
-            row.append(self.wait_spin)
-
-            self.wait_ms_label = Gtk.Label(label="ms")
-            row.append(self.wait_ms_label)
-
-            box.append(row)
-            self._update_options_visibility()
-        else:
-            self.rapidfire_check = None
-
-        return box
-
-    def _update_options_visibility(self):
-        if self.rapidfire_check:
-            rf_active = self.rapidfire_check.get_active()
-            self.hold_label.set_visible(rf_active)
-            self.hold_spin.set_visible(rf_active)
-            self.hold_ms_label.set_visible(rf_active)
-            self.wait_label.set_visible(rf_active)
-            self.wait_spin.set_visible(rf_active)
-            self.wait_ms_label.set_visible(rf_active)
-
-    def _on_tab_changed(self, stack, param):
-        child_name = self.stack.get_visible_child_name()
-        is_exec = child_name == "exec"
-        is_media = child_name == "media"
-        if self.rapidfire_check:
-            self.options_box.set_visible(not (is_exec or is_media))
-        self._update_actions_docs_button()
-
-    def _active_actions_docs_link(self) -> tuple[str, str] | None:
-        child_name = self.stack.get_visible_child_name()
-        if not child_name:
-            return None
-        return ACTION_DOC_LINKS.get(child_name)
-
-    def _update_actions_docs_button(self) -> None:
-        if not hasattr(self, "actions_docs_btn"):
-            return
-        link = self._active_actions_docs_link()
-        self.actions_docs_btn.set_visible(link is not None)
-        if link is None:
-            return
-        _anchor, title = link
-        self.actions_docs_btn.set_tooltip_text(f"Open {title} documentation")
-
-    def _on_actions_docs_clicked(self, _button: Gtk.Button) -> None:
-        link = self._active_actions_docs_link()
-        if link is None:
-            return
-        anchor, _title = link
-        url = _actions_docs_url(anchor)
-        try:
-            launcher = Gtk.UriLauncher.new(url)
-            launcher.launch(None, None, None)
-        except Exception:
-            log.exception("Could not open action documentation %s", url)
-
-    def _warn_and_clear_unsupported_rapidfire(self, action_type: ActionType) -> None:
-        if not self._rapidfire_enabled or action_type_supports_rapidfire(action_type):
-            return
-        log.warning(
-            "Ignoring rapidfire for unsupported %s action in superkey action dialog",
-            action_type.value,
-        )
-        if self.rapidfire_check and self.rapidfire_check.get_active():
-            self.rapidfire_check.set_active(False)
-        else:
-            self._rapidfire_enabled = False
-            self._update_options_visibility()
-
-    def _on_rapidfire_toggled(self, check):
-        self._rapidfire_enabled = check.get_active()
-        self._update_options_visibility()
-
-    def _build_exec_tab(self) -> Gtk.Widget:
-        box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=16)
-        box.set_margin_top(16)
-        box.set_margin_bottom(16)
-        box.set_margin_start(16)
-        box.set_margin_end(16)
-        box.set_valign(Gtk.Align.CENTER)
-
-        label = Gtk.Label(label="Execute a shell command:")
-        label.set_halign(Gtk.Align.START)
-        box.append(label)
-
-        self.cmd_entry = Gtk.Entry()
-        self.cmd_entry.set_placeholder_text("e.g., notify-send 'Hello'")
-        self.cmd_entry.set_hexpand(True)
-        if self._current_action and self._current_action.action_type == ActionType.EXEC:
-            self.cmd_entry.set_text(self._current_action.cmd or "")
-        box.append(self.cmd_entry)
-
-        exec_btn = Gtk.Button(label="Execute Command")
-        exec_btn.add_css_class("suggested-action")
-        exec_btn.connect("clicked", self._on_exec_clicked)
-        box.append(exec_btn)
-
-        return box
-
-    def _build_macro_tab(self) -> Gtk.Widget:
-        outer = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=8)
-        outer.set_margin_top(12)
-        outer.set_margin_bottom(12)
-        outer.set_margin_start(12)
-        outer.set_margin_end(12)
-
-        label = Gtk.Label(label="Trigger a saved macro")
-        label.add_css_class("dim-label")
-        label.set_halign(Gtk.Align.START)
-        outer.append(label)
-
-        self._superkey_macro_listbox = Gtk.ListBox()
-        self._superkey_macro_listbox.set_selection_mode(Gtk.SelectionMode.SINGLE)
-        self._superkey_macro_listbox.connect("row-selected", self._on_superkey_macro_selected)
-
-        scrolled = Gtk.ScrolledWindow()
-        scrolled.set_vexpand(True)
-        scrolled.set_child(self._superkey_macro_listbox)
-        outer.append(scrolled)
-
-        self._superkey_macro_map_btn = Gtk.Button(label="Map Macro")
-        self._superkey_macro_map_btn.add_css_class("suggested-action")
-        self._superkey_macro_map_btn.set_sensitive(False)
-        self._superkey_macro_map_btn.connect("clicked", self._on_superkey_macro_map_clicked)
-        outer.append(self._superkey_macro_map_btn)
-
-        return outer
-
-    def _load_superkey_macro_list(self) -> bool:
-        session_request_async({"command": "list_macros"}, self._on_superkey_macro_list_loaded)
-        return False
-
-    def _on_superkey_macro_list_loaded(self, result: dict | None) -> bool:
-        self._superkey_macro_list = (result or {}).get("macros", [])
-        self._populate_superkey_macro_listbox()
-        return False
-
-    def _populate_superkey_macro_listbox(self) -> None:
-        while self._superkey_macro_listbox.get_first_child():
-            self._superkey_macro_listbox.remove(self._superkey_macro_listbox.get_first_child())
-
-        if not self._superkey_macro_list:
-            row = Gtk.ListBoxRow()
-            row.set_selectable(False)
-            lbl = Gtk.Label(label="No macros saved yet")
-            lbl.add_css_class("dim-label")
-            lbl.set_margin_top(12)
-            lbl.set_margin_bottom(12)
-            row.set_child(lbl)
-            self._superkey_macro_listbox.append(row)
-            self._superkey_macro_map_btn.set_sensitive(False)
-            return
-
-        sel_row = None
-        for _i, macro in enumerate(self._superkey_macro_list):
-            row = Gtk.ListBoxRow()
-            box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-            box.set_margin_top(8)
-            box.set_margin_bottom(8)
-            box.set_margin_start(8)
-            box.set_margin_end(8)
-
-            name = macro.get("name", "")
-            lbl = Gtk.Label(label=name)
-            lbl.set_halign(Gtk.Align.START)
-            lbl.set_hexpand(True)
-            box.append(lbl)
-
-            count = int(macro.get("event_count", 0))
-            meta = Gtk.Label(label=f"{count} events")
-            meta.add_css_class("caption")
-            meta.add_css_class("dim-label")
-            box.append(meta)
-
-            row.set_child(box)
-            row._macro_name = name
-            self._superkey_macro_listbox.append(row)
-
-            if self._superkey_selected_macro and name == self._superkey_selected_macro:
-                sel_row = row
-
-        if sel_row is not None:
-            self._superkey_macro_listbox.select_row(sel_row)
-
-    def _on_superkey_macro_selected(self, list_box, row) -> None:
-        if row and hasattr(row, "_macro_name"):
-            self._superkey_selected_macro = row._macro_name
-        self._superkey_macro_map_btn.set_sensitive(bool(self._superkey_selected_macro))
-
-    def _on_superkey_macro_map_clicked(self, btn) -> None:
-        if not self._superkey_selected_macro:
-            return
-        self._warn_and_clear_unsupported_rapidfire(ActionType.MACRO)
-        action = SuperkeyAction(
-            action_type=ActionType.MACRO,
-            macro_name=self._superkey_selected_macro,
-            rapidfire_enabled=self._rapidfire_enabled if self.rapidfire_check else False,
-            rapidfire_hold_ms=int(self.hold_spin.get_value()) if self.rapidfire_check else 20,
-            rapidfire_wait_ms=int(self.wait_spin.get_value()) if self.rapidfire_check else 20,
-        )
-        self.emit("action-selected", action)
-        self.close()
-
-    def _on_exec_clicked(self, btn):
-        cmd = self.cmd_entry.get_text().strip()
-        if cmd:
-            self._warn_and_clear_unsupported_rapidfire(ActionType.EXEC)
-            action = SuperkeyAction(
-                action_type=ActionType.EXEC,
-                cmd=cmd,
-            )
-            self.emit("action-selected", action)
-            self.close()
-
-    def _on_mpris_clicked(self, btn, command: str):
-        self._warn_and_clear_unsupported_rapidfire(ActionType.MPRIS)
-        action = SuperkeyAction(
-            action_type=ActionType.MPRIS,
-            mpris_command=command,
-        )
-        self.emit("action-selected", action)
-        self.close()
-
-    def _on_clear_clicked(self, btn):
-        self.emit("action-selected", None)
-        self.close()
-
-    def _on_keyboard_clicked(self, btn, evdev_name: str):
-        use_rapidfire = bool(
-            self.rapidfire_check and _keyboard_target_allows_rapidfire(evdev_name)
-        )
-        action = SuperkeyAction(
-            action_type=ActionType.KEYBOARD,
-            target=evdev_name,
-            rapidfire_enabled=self._rapidfire_enabled if use_rapidfire else False,
-            rapidfire_hold_ms=int(self.hold_spin.get_value()) if use_rapidfire else 20,
-            rapidfire_wait_ms=int(self.wait_spin.get_value()) if use_rapidfire else 20,
-        )
-        self.emit("action-selected", action)
-        self.close()
-
-    def _on_mouse_clicked(self, btn, evdev_name: str):
-        action = SuperkeyAction(
-            action_type=ActionType.MOUSE,
-            target=evdev_name,
-            rapidfire_enabled=self._rapidfire_enabled if self.rapidfire_check else False,
-            rapidfire_hold_ms=int(self.hold_spin.get_value()) if self.rapidfire_check else 20,
-            rapidfire_wait_ms=int(self.wait_spin.get_value()) if self.rapidfire_check else 20,
-        )
-        self.emit("action-selected", action)
-        self.close()
-
-    def _on_gamepad_clicked(self, btn, evdev_name: str):
-        action = SuperkeyAction(
-            action_type=ActionType.GAMEPAD,
-            target=evdev_name,
-            output_id=self._selected_gamepad_output_id,
-            rapidfire_enabled=self._rapidfire_enabled if self.rapidfire_check else False,
-            rapidfire_hold_ms=int(self.hold_spin.get_value()) if self.rapidfire_check else 20,
-            rapidfire_wait_ms=int(self.wait_spin.get_value()) if self.rapidfire_check else 20,
-        )
-        self.emit("action-selected", action)
-        self.close()
-
-    def _on_gamepad_axis_clicked(self, btn, axis_target: str, axis_value: int):
-        action = SuperkeyAction(
-            action_type=ActionType.GAMEPAD_AXIS,
-            target=axis_target,
-            axis_value=int(axis_value),
-            output_id=self._selected_gamepad_output_id,
-            rapidfire_enabled=self._rapidfire_enabled if self.rapidfire_check else False,
-            rapidfire_hold_ms=int(self.hold_spin.get_value()) if self.rapidfire_check else 20,
-            rapidfire_wait_ms=int(self.wait_spin.get_value()) if self.rapidfire_check else 20,
-        )
-        self.emit("action-selected", action)
-        self.close()
-
-    def _on_gamepad_code_clicked(self, widget) -> None:
-        evdev_name = _resolve_gamepad_button_target(self.gamepad_code_entry.get_text())
-        if not evdev_name:
-            self.gamepad_code_entry.set_text("")
-            self.gamepad_code_entry.set_placeholder_text("Unknown button code")
-            return
-        self._on_gamepad_clicked(widget, evdev_name)
-
-    def _set_initial_tab(self):
-        if not self._current_action:
-            return
-        tab_map = {
-            ActionType.KEYBOARD: (
-                "media" if self._current_action.target in MEDIA_KEY_TARGETS else "keyboard"
-            ),
-            ActionType.MOUSE: "mouse",
-            ActionType.GAMEPAD: "gamepad",
-            ActionType.GAMEPAD_AXIS: "gamepad",
-            ActionType.MACRO: "macro",
-            ActionType.EXEC: "exec",
-            ActionType.MPRIS: "media",
-        }
-        name = tab_map.get(self._current_action.action_type)
-        if name:
-            self.stack.set_visible_child_name(name)
-
-    def _create_key_button(
-        self, label: str, evdev: str, width: float = 1, large: bool = False, protected: bool = False
-    ) -> Gtk.Button:
-        btn = Gtk.Button(label=label)
-        btn.add_css_class("key-button")
-
-        if large:
-            btn.set_size_request(200, 50)
-        else:
-            base_width = 36
-            btn.set_size_request(int(base_width * width), 34)
-
-        btn._evdev_name = evdev
-        return btn
-
-    def _build_keyboard_tab(self) -> Gtk.Widget:
-        return build_shared_keyboard_tab(
-            self,
-            keyboard_layout=KEYBOARD_LAYOUT,
-            key_to_evdev=KEY_TO_EVDEV,
-            key_widths=KEY_WIDTHS,
-            system_key_groups=SYSTEM_KEY_GROUPS,
-        )
-
-    def _build_navigation_tab(self) -> Gtk.Widget:
-        return build_shared_navigation_tab(self, f_extra=F_EXTRA)
-
-    def _build_media_tab(self) -> Gtk.Widget:
-        return build_shared_media_tab(
-            self,
-            media_groups=MEDIA_KEY_GROUPS,
-            mpris_groups=MPRIS_MEDIA_GROUPS,
-        )
-
-    def _build_mouse_tab(self) -> Gtk.Widget:
-        return build_shared_mouse_tab(self)
-
-    def _build_gamepad_tab(self) -> Gtk.Widget:
-        choices = self._gamepad_output_choices()
-        outer = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=8)
-        outer.set_margin_top(8)
-        if len(choices) > 1 or self._selected_gamepad_output_id:
-            row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-            row.set_margin_start(12)
-            row.set_margin_end(12)
-            label = Gtk.Label(label="Output")
-            label.set_xalign(0)
-            label.set_hexpand(True)
-            self._gamepad_output_ids = [output_id for output_id, _label in choices]
-            dropdown = Gtk.DropDown.new_from_strings([label for _output_id, label in choices])
-            selected = 0
-            for index, output_id in enumerate(self._gamepad_output_ids):
-                if gamepad_output_choice_matches(output_id, self._selected_gamepad_output_id):
-                    selected = index
-                    break
-            dropdown.set_selected(selected)
-            dropdown.connect("notify::selected", self._on_gamepad_output_selected)
-            row.append(label)
-            row.append(dropdown)
-            outer.append(row)
-        warning = Gtk.Label(label="")
-        warning.set_margin_start(12)
-        warning.set_margin_end(12)
-        warning.set_wrap(True)
-        warning.set_xalign(0)
-        warning.add_css_class("dim-label")
-        warning.add_css_class("warning")
-        self._gamepad_output_warning_label = warning
-        outer.append(warning)
-        outer.append(build_shared_gamepad_tab(self))
-        self._update_gamepad_output_warning()
-        self._prefill_gamepad_inputs()
-        return outer
-
-    def _gamepad_output_choices(self) -> list[tuple[str | None, str]]:
-        count = virtual_gamepad_count()
-        try:
-            hardware_configs = list(HardwareManager().list_hardware())
-        except (OSError, RuntimeError) as exc:
-            log.debug("Unable to load hardware configs for gamepad outputs: %s", exc)
-            hardware_configs = []
-        return gamepad_output_choices_for(
-            self._selected_gamepad_output_id,
-            count,
-            hardware_configs,
-        )
-
-    def _on_gamepad_output_selected(self, dropdown: Gtk.DropDown, _param) -> None:
-        selected = int(dropdown.get_selected())
-        if 0 <= selected < len(self._gamepad_output_ids):
-            self._selected_gamepad_output_id = self._gamepad_output_ids[selected]
-        self._update_gamepad_output_warning()
-
-    def _update_gamepad_output_warning(self) -> None:
-        label = self._gamepad_output_warning_label
-        if label is None:
-            return
-        message = gamepad_output_unavailable_message(
-            self._selected_gamepad_output_id,
-            virtual_gamepad_count(),
-        )
-        label.set_label(message or "")
-        label.set_visible(bool(message))
-
-    def _on_f_key_selected(self, btn):
-        idx = self.f_dropdown.get_selected()
-        f_key = F_EXTRA[idx]
-        evdev_name = KEY_TO_EVDEV.get(f_key)
-        if evdev_name:
-            self._on_keyboard_clicked(btn, evdev_name)
-
-    def _on_f_dropdown_changed(self, dropdown, pspec, btn: Gtk.Button):
-        idx = dropdown.get_selected()
-        f_key = F_EXTRA[idx]
-        btn.set_label(f"Map {f_key}")
